Move waveform display into a widget
