+++ src/wave_view_traces.h
@@ -1,0 +1,8 @@
+#ifndef WAVE_WAVE_VIEW_TRACES
+#define WAVE_WAVE_VIEW_TRACES
+
+#include <gtk/gtk.h>
+
+void rendertraces(cairo_t *cr);
+
+#endif
