+++ src/wavewindow.c
@@ -20,6 +20,7 @@
 #include "debug.h"
 #include "main.h"
 #include "signal_list.h"
+#include "wave_view.h"
 
 #if !defined _ISOC99_SOURCE
 #define _ISOC99_SOURCE 1
@@ -28,12 +29,9 @@
 
 #define WAVE_CAIRO_050_OFFSET (GLOBALS->cairo_050_offset)
 
-static void rendertimebar(void);
-static void draw_hptr_trace(Trptr t, hptr h, int which, int dodraw, int kill_grid);
-static void draw_hptr_trace_vector(Trptr t, hptr h, int which);
-static void draw_vptr_trace(Trptr t, vptr v, int which);
-static void rendertraces(void);
-static void rendertimes(void);
+static void rendertimebar(cairo_t *cr);
+static void rendertraces(cairo_t *cr);
+static void rendertimes(cairo_t *cr);
 
 static const GdkModifierType bmask[4] = {0,
                                          GDK_BUTTON1_MASK,
@@ -46,29 +44,7 @@ static const GdkEventMask m_bmask[4] = {0,
                                         GDK_BUTTON3_MOTION_MASK}; /* button 1, 3 motion encodings */
 #endif
 
-void wavewindow_paint(cairo_t *cr)
-{
-    int scale_factor = gtk_widget_get_scale_factor(GLOBALS->signalarea);
-    cairo_matrix_t prev_matrix;
-    cairo_get_matrix(cr, &prev_matrix);
-    cairo_scale(cr, 1.0 / scale_factor, 1.0 / scale_factor);
-    cairo_set_source_surface(cr, GLOBALS->surface_wavepixmap_wavewindow_c_1, 0, 0);
-    cairo_paint(cr);
-    cairo_set_matrix(cr, &prev_matrix);
-}
-
 /******************************************************************/
-
-static void update_dual(void)
-{
-    if (GLOBALS->dual_ctx && !GLOBALS->dual_race_lock) {
-        GLOBALS->dual_ctx[GLOBALS->dual_id].zoom = GLOBALS->tims.zoom;
-        GLOBALS->dual_ctx[GLOBALS->dual_id].marker = GLOBALS->tims.marker;
-        GLOBALS->dual_ctx[GLOBALS->dual_id].baseline = GLOBALS->tims.baseline;
-        GLOBALS->dual_ctx[GLOBALS->dual_id].left_margin_time = GLOBALS->tims.start;
-        GLOBALS->dual_ctx[GLOBALS->dual_id].use_new_times = 1;
-    }
-}
 
 /******************************************************************/
 static int gesture_filter_set = 0; /* to prevent floods of gesture events before next draw */
@@ -440,7 +416,7 @@ xit:
     return (marker);
 }
 
-static void render_individual_named_marker(int i, wave_rgb_t gc, int blackout)
+static void render_individual_named_marker(cairo_t *cr, int i, wave_rgb_t gc, int blackout)
 {
     gdouble pixstep;
     gint xl;
@@ -459,21 +435,13 @@ static void render_individual_named_marker(int i, wave_rgb_t gc, int blackout)
                 char nbuff[16];
                 make_bijective_marker_id_string(nbuff, i);
 
-                cairo_set_dash(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                               dashed1,
-                               sizeof(dashed1) / sizeof(dashed1[0]),
-                               0);
-                XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                  gc,
-                                  xl,
-                                  GLOBALS->fontheight - 1,
-                                  xl,
-                                  GLOBALS->waveheight - 1);
-                cairo_set_dash(GLOBALS->cr_wavepixmap_wavewindow_c_1, dashed1, 0, 0);
+                cairo_set_dash(cr, dashed1, sizeof(dashed1) / sizeof(dashed1[0]), 0);
+                XXX_gdk_draw_line(cr, gc, xl, GLOBALS->fontheight - 1, xl, GLOBALS->waveheight - 1);
+                cairo_set_dash(cr, dashed1, 0, 0);
 
                 if ((!GLOBALS->marker_names[i]) || (!GLOBALS->marker_names[i][0])) {
                     XXX_font_engine_draw_string(
-                        GLOBALS->cr_wavepixmap_wavewindow_c_1,
+                        cr,
                         GLOBALS->wavefont_smaller,
                         &gc,
                         xl - (font_engine_string_measure(GLOBALS->wavefont_smaller, nbuff) >> 1) +
@@ -487,7 +455,7 @@ static void render_individual_named_marker(int i, wave_rgb_t gc, int blackout)
                                      marker labels */
                     {
                         XXX_gdk_draw_rectangle(
-                            GLOBALS->cr_wavepixmap_wavewindow_c_1,
+                            cr,
                             GLOBALS->rgb_gc.gc_timeb_wavewindow_c_1,
                             TRUE,
                             xl - (width >> 1),
@@ -496,7 +464,7 @@ static void render_individual_named_marker(int i, wave_rgb_t gc, int blackout)
                             GLOBALS->wavefont_smaller->ascent + GLOBALS->wavefont_smaller->descent);
                     }
 
-                    XXX_font_engine_draw_string(GLOBALS->cr_wavepixmap_wavewindow_c_1,
+                    XXX_font_engine_draw_string(cr,
                                                 GLOBALS->wavefont_smaller,
                                                 &gc,
                                                 xl - (width >> 1) + WAVE_CAIRO_050_OFFSET,
@@ -508,21 +476,19 @@ static void render_individual_named_marker(int i, wave_rgb_t gc, int blackout)
     }
 }
 
-static void draw_named_markers(void)
+static void draw_named_markers(cairo_t *cr)
 {
     int i;
 
-    if (!GLOBALS->cr_wavepixmap_wavewindow_c_1)
-        return;
-
     for (i = 0; i < WAVE_NUM_NAMED_MARKERS; i++) {
         if (i != GLOBALS->named_marker_lock_idx) {
-            render_individual_named_marker(i, GLOBALS->rgb_gc.gc_mark_wavewindow_c_1, 0);
+            render_individual_named_marker(cr, i, GLOBALS->rgb_gc.gc_mark_wavewindow_c_1, 0);
         }
     }
 
     if (GLOBALS->named_marker_lock_idx >= 0) {
-        render_individual_named_marker(GLOBALS->named_marker_lock_idx,
+        render_individual_named_marker(cr,
+                                       GLOBALS->named_marker_lock_idx,
                                        GLOBALS->rgb_gc.gc_umark_wavewindow_c_1,
                                        1);
     }
@@ -544,131 +510,31 @@ static void sync_marker(void)
     }
 }
 
-static void draw_marker(void)
+static void draw_marker_partitions(cairo_t *cr)
 {
-    gdouble pixstep;
-    gint xl;
+    (void)cr;
 
-    if (!gtk_widget_get_window(GLOBALS->wavearea))
-        return;
+    g_printerr("draw_marker_partitions disabled\n");
+    //     draw_marker();
 
-#ifdef WAVE_ALLOW_GTK3_CAIRO_CREATE_FIX
-    GdkDrawingContext *gdc;
-#endif
-    cairo_t *cr =
-        XXX_gdk_cairo_create(XXX_GDK_DRAWABLE(gtk_widget_get_window(GLOBALS->wavearea)), &gdc);
-    cairo_set_line_width(cr, GLOBALS->cr_line_width);
-    cairo_set_line_cap(cr, CAIRO_LINE_CAP_SQUARE);
+    // #ifdef WAVE_ALLOW_GTK3_CAIRO_CREATE_FIX
+    //     GdkDrawingContext *gdc;
+    // #endif
+    //     cairo_t *cr =
+    //         XXX_gdk_cairo_create(XXX_GDK_DRAWABLE(gtk_widget_get_window(GLOBALS->wavearea)),
+    //         &gdc);
+    //     wavewindow_paint(cr);
 
-    GLOBALS->m1x_wavewindow_c_1 = GLOBALS->m2x_wavewindow_c_1 = -1;
+    // #ifdef WAVE_ALLOW_GTK3_CAIRO_CREATE_FIX
+    //     gdk_window_end_draw_frame(gtk_widget_get_window(GLOBALS->wavearea), gdc);
+    // #else
+    //     cairo_destroy(cr);
+    // #endif
 
-    if (GLOBALS->tims.baseline >= 0) {
-        if ((GLOBALS->tims.baseline >= GLOBALS->tims.start) &&
-            (GLOBALS->tims.baseline <= GLOBALS->tims.last) &&
-            (GLOBALS->tims.baseline <= GLOBALS->tims.end)) {
-            pixstep = ((gdouble)GLOBALS->nsperframe) / ((gdouble)GLOBALS->pixelsperframe);
-            xl = ((gdouble)(GLOBALS->tims.baseline - GLOBALS->tims.start)) /
-                 pixstep; /* snap to integer */
-            if ((xl >= 0) && (xl < GLOBALS->wavewidth)) {
-                XXX_gdk_draw_line(cr,
-                                  GLOBALS->rgb_gc.gc_baseline_wavewindow_c_1,
-                                  xl,
-                                  GLOBALS->fontheight - 1,
-                                  xl,
-                                  GLOBALS->waveheight - 1);
-            }
-        }
-    }
-
-    if (GLOBALS->tims.marker >= 0) {
-        if ((GLOBALS->tims.marker >= GLOBALS->tims.start) &&
-            (GLOBALS->tims.marker <= GLOBALS->tims.last) &&
-            (GLOBALS->tims.marker <= GLOBALS->tims.end)) {
-            pixstep = ((gdouble)GLOBALS->nsperframe) / ((gdouble)GLOBALS->pixelsperframe);
-            xl = ((gdouble)(GLOBALS->tims.marker - GLOBALS->tims.start)) /
-                 pixstep; /* snap to integer */
-            if ((xl >= 0) && (xl < GLOBALS->wavewidth)) {
-                XXX_gdk_draw_line(cr,
-                                  GLOBALS->rgb_gc.gc_umark_wavewindow_c_1,
-                                  xl,
-                                  GLOBALS->fontheight - 1,
-                                  xl,
-                                  GLOBALS->waveheight - 1);
-                GLOBALS->m1x_wavewindow_c_1 = xl;
-            }
-        }
-    }
-
-    if ((GLOBALS->enable_ghost_marker) && (GLOBALS->in_button_press_wavewindow_c_1) &&
-        (GLOBALS->tims.lmbcache >= 0)) {
-        if ((GLOBALS->tims.lmbcache >= GLOBALS->tims.start) &&
-            (GLOBALS->tims.lmbcache <= GLOBALS->tims.last) &&
-            (GLOBALS->tims.lmbcache <= GLOBALS->tims.end)) {
-            pixstep = ((gdouble)GLOBALS->nsperframe) / ((gdouble)GLOBALS->pixelsperframe);
-            xl = ((gdouble)(GLOBALS->tims.lmbcache - GLOBALS->tims.start)) /
-                 pixstep; /* snap to integer */
-            if ((xl >= 0) && (xl < GLOBALS->wavewidth)) {
-                XXX_gdk_draw_line(cr,
-                                  GLOBALS->rgb_gc.gc_umark_wavewindow_c_1,
-                                  xl,
-                                  GLOBALS->fontheight - 1,
-                                  xl,
-                                  GLOBALS->waveheight - 1);
-                GLOBALS->m2x_wavewindow_c_1 = xl;
-            }
-        }
-    }
-
-    if (GLOBALS->m1x_wavewindow_c_1 >
-        GLOBALS->m2x_wavewindow_c_1) /* ensure m1x <= m2x for partitioned refresh */
-    {
-        gint t;
-
-        t = GLOBALS->m1x_wavewindow_c_1;
-        GLOBALS->m1x_wavewindow_c_1 = GLOBALS->m2x_wavewindow_c_1;
-        GLOBALS->m2x_wavewindow_c_1 = t;
-    }
-
-    if (GLOBALS->m1x_wavewindow_c_1 == -1)
-        GLOBALS->m1x_wavewindow_c_1 =
-            GLOBALS->m2x_wavewindow_c_1; /* make both markers same if no ghost marker or v.v. */
-
-    update_dual();
-
-#ifdef WAVE_ALLOW_GTK3_CAIRO_CREATE_FIX
-    gdk_window_end_draw_frame(gtk_widget_get_window(GLOBALS->wavearea), gdc);
-#else
-    cairo_destroy(cr);
-#endif
-
-    /*
-    #ifdef GDK_WINDOWING_WAYLAND
-    if(GDK_IS_WAYLAND_DISPLAY(gdk_display_get_default())) gtk_widget_queue_draw(GLOBALS->wavearea);
-    #endif
-    */
+    //     draw_marker();
 }
 
-static void draw_marker_partitions(void)
-{
-    draw_marker();
-
-#ifdef WAVE_ALLOW_GTK3_CAIRO_CREATE_FIX
-    GdkDrawingContext *gdc;
-#endif
-    cairo_t *cr =
-        XXX_gdk_cairo_create(XXX_GDK_DRAWABLE(gtk_widget_get_window(GLOBALS->wavearea)), &gdc);
-    wavewindow_paint(cr);
-
-#ifdef WAVE_ALLOW_GTK3_CAIRO_CREATE_FIX
-    gdk_window_end_draw_frame(gtk_widget_get_window(GLOBALS->wavearea), gdc);
-#else
-    cairo_destroy(cr);
-#endif
-
-    draw_marker();
-}
-
-static void renderblackout(void)
+static void renderblackout(cairo_t *cr)
 {
     gfloat pageinc;
     TimeType lhs, rhs, lclip, rclip;
@@ -699,7 +565,7 @@ static void renderblackout(void)
                 if (rclip > ((GLOBALS->wavewidth + 1) * GLOBALS->nspx))
                     rclip = (GLOBALS->wavewidth + 1) * (GLOBALS->nspx);
 
-                XXX_gdk_draw_rectangle(GLOBALS->cr_wavepixmap_wavewindow_c_1,
+                XXX_gdk_draw_rectangle(cr,
                                        GLOBALS->rgb_gc.gc_xfill_wavewindow_c_1,
                                        TRUE,
                                        (((gdouble)lclip) * GLOBALS->pxns),
@@ -720,47 +586,42 @@ static void service_hslider(GtkWidget *text, gpointer data)
 
     DEBUG(printf("Wave HSlider Moved\n"));
 
-    if ((GLOBALS->cr_wavepixmap_wavewindow_c_1) && (GLOBALS->wavewidth > 1)) {
-#ifdef WAVE_ALLOW_GTK3_GESTURE_EVENT
-        TimeType old_start = GLOBALS->tims.start;
-#endif
-        GtkAdjustment *hadj = GTK_ADJUSTMENT(GLOBALS->wave_hslider);
-
-        if (!GLOBALS->tims.timecache) {
-            GLOBALS->tims.start = time_trunc(gtk_adjustment_get_value(hadj));
-        } else {
-            GLOBALS->tims.start = time_trunc(GLOBALS->tims.timecache);
-            GLOBALS->tims.timecache = 0; /* reset */
-        }
-
-        if (GLOBALS->tims.start < GLOBALS->tims.first)
-            GLOBALS->tims.start = GLOBALS->tims.first;
-        else if (GLOBALS->tims.start > GLOBALS->tims.last)
-            GLOBALS->tims.start = GLOBALS->tims.last;
-
-        GLOBALS->tims.laststart = GLOBALS->tims.start;
-
-#ifdef WAVE_ALLOW_GTK3_GESTURE_EVENT
-        if ((gesture_in_zoom) || (!GLOBALS->swipe_init_time) ||
-            (GLOBALS->wavearea_gesture_swipe_velocity_x == 0.0) ||
-            (old_start != GLOBALS->tims.start)) /* cut down on redundant draws when swiping */
-#endif
-        {
-            XXX_gdk_draw_rectangle(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                   GLOBALS->rgb_gc.gc_back_wavewindow_c_1,
-                                   TRUE,
-                                   0,
-                                   0,
-                                   GLOBALS->wavewidth,
-                                   GLOBALS->waveheight);
-            rendertimebar();
-            gtk_widget_queue_draw(GLOBALS->wavewindow);
-        }
-#ifdef WAVE_ALLOW_GTK3_GESTURE_EVENT
-        if (gesture_in_zoom)
-            gesture_in_zoom--;
-#endif
+    if (GLOBALS->wavewidth < 1) {
+        return;
     }
+
+#ifdef WAVE_ALLOW_GTK3_GESTURE_EVENT
+    TimeType old_start = GLOBALS->tims.start;
+#endif
+    GtkAdjustment *hadj = GTK_ADJUSTMENT(GLOBALS->wave_hslider);
+
+    if (!GLOBALS->tims.timecache) {
+        GLOBALS->tims.start = time_trunc(gtk_adjustment_get_value(hadj));
+    } else {
+        GLOBALS->tims.start = time_trunc(GLOBALS->tims.timecache);
+        GLOBALS->tims.timecache = 0; /* reset */
+    }
+
+    if (GLOBALS->tims.start < GLOBALS->tims.first)
+        GLOBALS->tims.start = GLOBALS->tims.first;
+    else if (GLOBALS->tims.start > GLOBALS->tims.last)
+        GLOBALS->tims.start = GLOBALS->tims.last;
+
+    GLOBALS->tims.laststart = GLOBALS->tims.start;
+
+#ifdef WAVE_ALLOW_GTK3_GESTURE_EVENT
+    if ((gesture_in_zoom) || (!GLOBALS->swipe_init_time) ||
+        (GLOBALS->wavearea_gesture_swipe_velocity_x == 0.0) ||
+        (old_start != GLOBALS->tims.start)) /* cut down on redundant draws when swiping
+                                             */
+#endif
+    {
+        gw_wave_view_force_redraw(GW_WAVE_VIEW(GLOBALS->wavearea));
+    }
+#ifdef WAVE_ALLOW_GTK3_GESTURE_EVENT
+    if (gesture_in_zoom)
+        gesture_in_zoom--;
+#endif
 }
 
 #ifndef WAVE_GTK3_SIZE_ALLOCATE_WORKAROUND_WAVE_VSLIDER
@@ -773,24 +634,9 @@ static
 
     sync_marker();
 
-    if (GLOBALS->cr_wavepixmap_wavewindow_c_1) {
-        XXX_gdk_draw_rectangle(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                               GLOBALS->rgb_gc.gc_back_wavewindow_c_1,
-                               TRUE,
-                               0,
-                               0,
-                               GLOBALS->wavewidth,
-                               GLOBALS->waveheight);
-
-        /* if(GLOBALS->display_grid) */ rendertimes();
-
-        rendertraces();
-    }
+    gw_wave_view_force_redraw(GW_WAVE_VIEW(GLOBALS->wavearea));
 
     GLOBALS->old_wvalue = gtk_adjustment_get_value(sadj);
-
-    draw_named_markers();
-    draw_marker();
 }
 
 void button_press_release_common(void)
@@ -848,7 +694,7 @@ static void button_motion_common(gint xin, gint yin, int pressrel, int is_button
         if (GLOBALS->tims.lmbcache < 0)
             GLOBALS->tims.lmbcache = time_trunc(newcurr);
 
-        draw_marker_partitions();
+        // draw_marker_partitions(cr);
 
         if ((pressrel) || (GLOBALS->constant_marker_update)) {
             button_press_release_common();
@@ -859,8 +705,10 @@ static void button_motion_common(gint xin, gint yin, int pressrel, int is_button
                                      : -1;
         update_basetime(GLOBALS->tims.baseline);
         update_markertime(GLOBALS->tims.marker);
-        wavearea_configure_event(GLOBALS->wavearea, NULL);
+        // gw_wave_view_force_redraw(GW_WAVE_VIEW(GLOBALS->wavearea));
     }
+
+    gtk_widget_queue_draw(GLOBALS->wavearea);
 }
 
 static gint motion_notify_event(GtkWidget *widget, GdkEventMotion *event)
@@ -1013,7 +861,7 @@ static gint motion_notify_event(GtkWidget *widget, GdkEventMotion *event)
                 ...commented out to reduce on visual noise */
 
                 gw_signal_list_force_redraw(GW_SIGNAL_LIST(GLOBALS->signalarea));
-                wavearea_configure_event(GLOBALS->wavearea, NULL);
+                gw_wave_view_force_redraw(GW_WAVE_VIEW(GLOBALS->wavearea));
             }
         }
 
@@ -1645,139 +1493,6 @@ void make_sigarea_gcs(GtkWidget *signalarea)
     }
 }
 
-static const int wave_rgb_rainbow[WAVE_NUM_RAINBOW] = WAVE_RAINBOW_RGB;
-
-gint wavearea_configure_event(GtkWidget *widget, GdkEventConfigure *event)
-{
-    (void)event;
-
-    if ((!widget) || (!gtk_widget_get_window(widget)))
-        return (TRUE);
-
-    GtkAllocation allocation;
-    gtk_widget_get_allocation(widget, &allocation);
-
-    int scale_factor = gtk_widget_get_scale_factor(widget);
-
-    DEBUG(printf("WaveWin Configure Event h: %d, w: %d\n", allocation.height, allocation.width));
-
-    if (GLOBALS->cr_wavepixmap_wavewindow_c_1) {
-        if ((GLOBALS->wavewidth != allocation.width) ||
-            (GLOBALS->waveheight != allocation.height)) {
-            GLOBALS->wavewidth = allocation.width;
-            GLOBALS->waveheight = allocation.height;
-
-            cairo_destroy(GLOBALS->cr_wavepixmap_wavewindow_c_1);
-            cairo_surface_destroy(GLOBALS->surface_wavepixmap_wavewindow_c_1);
-
-            GLOBALS->surface_wavepixmap_wavewindow_c_1 =
-                cairo_image_surface_create(CAIRO_FORMAT_RGB24,
-                                           GLOBALS->wavewidth * scale_factor,
-                                           allocation.height * scale_factor);
-            GLOBALS->cr_wavepixmap_wavewindow_c_1 =
-                cairo_create(GLOBALS->surface_wavepixmap_wavewindow_c_1);
-            cairo_scale(GLOBALS->cr_wavepixmap_wavewindow_c_1, scale_factor, scale_factor);
-            cairo_set_line_width(GLOBALS->cr_wavepixmap_wavewindow_c_1, GLOBALS->cr_line_width);
-            cairo_set_line_cap(GLOBALS->cr_wavepixmap_wavewindow_c_1, CAIRO_LINE_CAP_SQUARE);
-        }
-        GLOBALS->old_wvalue = -1.0;
-    } else {
-        GLOBALS->wavewidth = allocation.width;
-        GLOBALS->waveheight = allocation.height;
-
-        GLOBALS->surface_wavepixmap_wavewindow_c_1 =
-            cairo_image_surface_create(CAIRO_FORMAT_RGB24,
-                                       GLOBALS->wavewidth * scale_factor,
-                                       allocation.height * scale_factor);
-        GLOBALS->cr_wavepixmap_wavewindow_c_1 =
-            cairo_create(GLOBALS->surface_wavepixmap_wavewindow_c_1);
-        cairo_scale(GLOBALS->cr_wavepixmap_wavewindow_c_1, scale_factor, scale_factor);
-        cairo_set_line_width(GLOBALS->cr_wavepixmap_wavewindow_c_1, GLOBALS->cr_line_width);
-        cairo_set_line_cap(GLOBALS->cr_wavepixmap_wavewindow_c_1, CAIRO_LINE_CAP_SQUARE);
-    }
-
-    if (!GLOBALS->made_gc_contexts_wavewindow_c_1) {
-        int i;
-
-        GLOBALS->rgb_gc.gc_back_wavewindow_c_1 = XXX_alloc_color(GLOBALS->color_back);
-        GLOBALS->rgb_gc.gc_baseline_wavewindow_c_1 = XXX_alloc_color(GLOBALS->color_baseline);
-        GLOBALS->rgb_gc.gc_grid_wavewindow_c_1 = XXX_alloc_color(GLOBALS->color_grid);
-        GLOBALS->rgb_gc.gc_grid2_wavewindow_c_1 = XXX_alloc_color(GLOBALS->color_grid2);
-        GLOBALS->rgb_gc.gc_time_wavewindow_c_1 = XXX_alloc_color(GLOBALS->color_time);
-        GLOBALS->rgb_gc.gc_timeb_wavewindow_c_1 = XXX_alloc_color(GLOBALS->color_timeb);
-        GLOBALS->rgb_gc.gc_value_wavewindow_c_1 = XXX_alloc_color(GLOBALS->color_value);
-        GLOBALS->rgb_gc.gc_low_wavewindow_c_1 = XXX_alloc_color(GLOBALS->color_low);
-        GLOBALS->rgb_gc.gc_highfill_wavewindow_c_1 = XXX_alloc_color(GLOBALS->color_highfill);
-        GLOBALS->rgb_gc.gc_high_wavewindow_c_1 = XXX_alloc_color(GLOBALS->color_high);
-        GLOBALS->rgb_gc.gc_trans_wavewindow_c_1 = XXX_alloc_color(GLOBALS->color_trans);
-        GLOBALS->rgb_gc.gc_mid_wavewindow_c_1 = XXX_alloc_color(GLOBALS->color_mid);
-        GLOBALS->rgb_gc.gc_xfill_wavewindow_c_1 = XXX_alloc_color(GLOBALS->color_xfill);
-        GLOBALS->rgb_gc.gc_x_wavewindow_c_1 = XXX_alloc_color(GLOBALS->color_x);
-        GLOBALS->rgb_gc.gc_vbox_wavewindow_c_1 = XXX_alloc_color(GLOBALS->color_vbox);
-        GLOBALS->rgb_gc.gc_vtrans_wavewindow_c_1 = XXX_alloc_color(GLOBALS->color_vtrans);
-        GLOBALS->rgb_gc.gc_mark_wavewindow_c_1 = XXX_alloc_color(GLOBALS->color_mark);
-        GLOBALS->rgb_gc.gc_umark_wavewindow_c_1 = XXX_alloc_color(GLOBALS->color_umark);
-
-        GLOBALS->rgb_gc.gc_0_wavewindow_c_1 = XXX_alloc_color(GLOBALS->color_0);
-        GLOBALS->rgb_gc.gc_1fill_wavewindow_c_1 = XXX_alloc_color(GLOBALS->color_1fill);
-        GLOBALS->rgb_gc.gc_1_wavewindow_c_1 = XXX_alloc_color(GLOBALS->color_1);
-        GLOBALS->rgb_gc.gc_ufill_wavewindow_c_1 = XXX_alloc_color(GLOBALS->color_ufill);
-        GLOBALS->rgb_gc.gc_u_wavewindow_c_1 = XXX_alloc_color(GLOBALS->color_u);
-        GLOBALS->rgb_gc.gc_wfill_wavewindow_c_1 = XXX_alloc_color(GLOBALS->color_wfill);
-        GLOBALS->rgb_gc.gc_w_wavewindow_c_1 = XXX_alloc_color(GLOBALS->color_w);
-        GLOBALS->rgb_gc.gc_dashfill_wavewindow_c_1 = XXX_alloc_color(GLOBALS->color_dashfill);
-        GLOBALS->rgb_gc.gc_dash_wavewindow_c_1 = XXX_alloc_color(GLOBALS->color_dash);
-
-        GLOBALS->made_gc_contexts_wavewindow_c_1 = ~0;
-
-        memcpy(&GLOBALS->rgb_gccache, &GLOBALS->rgb_gc, sizeof(struct wave_rgbmaster_t));
-
-        /* add rainbow colors */
-        for (i = 0; i < WAVE_NUM_RAINBOW; i++) {
-            int col = wave_rgb_rainbow[i];
-
-            GLOBALS->rgb_gc_rainbow[i * 2] = XXX_alloc_color(col);
-            col >>= 1;
-            col &= 0x007F7F7F;
-            GLOBALS->rgb_gc_rainbow[i * 2 + 1] = XXX_alloc_color(col);
-        }
-    }
-
-    if (GLOBALS->timestart_from_savefile_valid) {
-        gfloat pageinc = (gfloat)(((gdouble)GLOBALS->wavewidth) * GLOBALS->nspx);
-
-        if ((GLOBALS->timestart_from_savefile >= GLOBALS->tims.first) &&
-            (GLOBALS->timestart_from_savefile <= (GLOBALS->tims.last - pageinc))) {
-            GtkAdjustment *hadj = GTK_ADJUSTMENT(GLOBALS->wave_hslider);
-            gtk_adjustment_set_value(hadj, (gdouble)(GLOBALS->timestart_from_savefile));
-            fix_wavehadj();
-            g_signal_emit_by_name(GTK_ADJUSTMENT(GLOBALS->wave_hslider),
-                                  "value_changed"); /* force zoom update */
-            g_signal_emit_by_name(GTK_ADJUSTMENT(GLOBALS->wave_hslider),
-                                  "changed"); /* force zoom update */
-        }
-        GLOBALS->timestart_from_savefile_valid--; /* for previous gtk toolkit versions could set
-                                                     this directly to zero */
-    }
-
-    if (GLOBALS->wavewidth > 1) {
-        if ((!GLOBALS->do_initial_zoom_fit) || (GLOBALS->do_initial_zoom_fit_used)) {
-            calczoom(GLOBALS->tims.zoom);
-            fix_wavehadj();
-            g_signal_emit_by_name(GTK_ADJUSTMENT(GLOBALS->wave_hslider),
-                                  "value_changed"); /* force zoom update */
-            g_signal_emit_by_name(GTK_ADJUSTMENT(GLOBALS->wave_hslider),
-                                  "changed"); /* force zoom update */
-        } else {
-            GLOBALS->do_initial_zoom_fit_used = 1;
-            service_zoom_fit(NULL, NULL);
-        }
-    }
-
-    /* tims.timecache=tims.laststart; */
-    return (TRUE);
-}
-
 /*
  * screengrab vs normal rendering gcs...
  */
@@ -1827,45 +1542,6 @@ void force_normal_gcs(void)
     GLOBALS->black_and_white = 0;
 
     memcpy(&GLOBALS->rgb_gc, &GLOBALS->rgb_gccache, sizeof(struct wave_rgbmaster_t));
-}
-
-static gint wavearea_configure_event_local(GtkWidget *widget, GdkEventConfigure *event)
-{
-    gint rc;
-    gint page_num = gtk_notebook_get_current_page(GTK_NOTEBOOK(GLOBALS->notebook));
-    struct Global *g_old = GLOBALS;
-
-    set_GLOBALS((*GLOBALS->contexts)[page_num]);
-
-    rc = wavearea_configure_event(widget, event);
-
-    set_GLOBALS(g_old);
-
-    return (rc);
-}
-
-static gint draw_event(GtkWidget *widget, cairo_t *cr, gpointer user_data)
-{
-    (void)widget;
-    (void)user_data;
-
-    gint rc = FALSE;
-    gint page_num = gtk_notebook_get_current_page(GTK_NOTEBOOK(GLOBALS->notebook));
-    /* struct Global *g_old = GLOBALS; */
-
-    set_GLOBALS((*GLOBALS->contexts)[page_num]);
-
-    wavewindow_paint(cr);
-
-    draw_marker();
-
-    /* seems to cause a conflict flipping back so don't! */
-    /* set_GLOBALS(g_old); */
-
-    if (gesture_filter_set)
-        gesture_filter_set = 0;
-
-    return (rc);
 }
 
 #ifdef WAVE_GTK3_SIZE_ALLOCATE_WORKAROUND_WAVE_VSLIDER
@@ -2303,24 +1979,9 @@ GtkWidget *create_wavewindow(void)
     gtk_container_set_resize_mode(GTK_CONTAINER(table), GTK_RESIZE_IMMEDIATE);
 #endif
 
-    GLOBALS->wavearea = gtk_drawing_area_new();
+    GLOBALS->wavearea = gw_wave_view_new();
     gtk_widget_set_hexpand(GLOBALS->wavearea, TRUE);
     gtk_widget_set_vexpand(GLOBALS->wavearea, TRUE);
-
-    gtk_widget_set_events(GLOBALS->wavearea,
-                          GDK_SCROLL_MASK | GDK_EXPOSURE_MASK | GDK_BUTTON_PRESS_MASK |
-                              GDK_BUTTON_RELEASE_MASK | GDK_POINTER_MOTION_MASK |
-                              GDK_POINTER_MOTION_HINT_MASK
-#ifdef WAVE_ALLOW_GTK3_GESTURE_EVENT
-                              | GDK_TOUCH_MASK | GDK_TABLET_PAD_MASK | GDK_TOUCHPAD_GESTURE_MASK
-#endif
-    );
-
-    g_signal_connect(GLOBALS->wavearea,
-                     "configure_event",
-                     G_CALLBACK(wavearea_configure_event_local),
-                     NULL);
-    g_signal_connect(GLOBALS->wavearea, "draw", G_CALLBACK(draw_event), NULL);
 
 #ifdef WAVE_ALLOW_GTK3_GESTURE_EVENT
     if (GLOBALS->use_gestures < 0) /* <0 means "maybe (if available, enable)" */
@@ -2994,2854 +2655,4 @@ void calczoom(gdouble z0)
                  z0,
                  (int)GLOBALS->pixelsperframe,
                  (float)GLOBALS->nsperframe));
-}
-
-static void renderhash(int x, TimeType tim)
-{
-    TimeType rborder;
-    int fhminus2;
-    int rhs;
-    gdouble dx;
-    gdouble hashoffset;
-    int iter = 0;
-    int s_ctx_iter;
-    int timearray_encountered = (GLOBALS->ruler_step != 0);
-
-    fhminus2 = GLOBALS->fontheight - 2;
-
-    WAVE_STRACE_ITERATOR(s_ctx_iter)
-    {
-        GLOBALS->strace_ctx = &GLOBALS->strace_windows[GLOBALS->strace_current_window = s_ctx_iter];
-        if (GLOBALS->strace_ctx->timearray) {
-            timearray_encountered = 1;
-            break;
-        }
-    }
-
-    cairo_set_source_rgba(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                          GLOBALS->rgb_gc.gc_grid_wavewindow_c_1.r,
-                          GLOBALS->rgb_gc.gc_grid_wavewindow_c_1.g,
-                          GLOBALS->rgb_gc.gc_grid_wavewindow_c_1.b,
-                          GLOBALS->rgb_gc.gc_grid_wavewindow_c_1.a);
-    cairo_move_to(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                  x + WAVE_CAIRO_050_OFFSET,
-                  WAVE_CAIRO_050_OFFSET);
-    cairo_line_to(
-        GLOBALS->cr_wavepixmap_wavewindow_c_1,
-        x + WAVE_CAIRO_050_OFFSET,
-        (((!timearray_encountered) && (GLOBALS->display_grid) && (GLOBALS->enable_vert_grid))
-             ? GLOBALS->waveheight
-             : fhminus2) +
-            WAVE_CAIRO_050_OFFSET);
-    cairo_stroke(GLOBALS->cr_wavepixmap_wavewindow_c_1);
-
-    if (tim == GLOBALS->tims.last)
-        return;
-
-    rborder = (GLOBALS->tims.last - GLOBALS->tims.start) * GLOBALS->pxns;
-    DEBUG(printf("Rborder: %lld, Wavewidth: %d\n", rborder, GLOBALS->wavewidth));
-    if (rborder > GLOBALS->wavewidth)
-        rborder = GLOBALS->wavewidth;
-    if ((rhs = x + GLOBALS->pixelsperframe) > rborder)
-        rhs = rborder;
-
-    cairo_move_to(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                  x + WAVE_CAIRO_050_OFFSET,
-                  GLOBALS->wavecrosspiece + WAVE_CAIRO_050_OFFSET);
-    cairo_line_to(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                  rhs + WAVE_CAIRO_050_OFFSET,
-                  GLOBALS->wavecrosspiece + WAVE_CAIRO_050_OFFSET);
-    cairo_stroke(GLOBALS->cr_wavepixmap_wavewindow_c_1);
-
-    dx = x + (hashoffset = GLOBALS->hashstep);
-    x = dx;
-
-    while ((hashoffset < GLOBALS->pixelsperframe) && (x <= rhs) && (iter < 9)) {
-        cairo_move_to(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                      x + WAVE_CAIRO_050_OFFSET,
-                      GLOBALS->wavecrosspiece + WAVE_CAIRO_050_OFFSET);
-        cairo_line_to(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                      x + WAVE_CAIRO_050_OFFSET,
-                      fhminus2 + WAVE_CAIRO_050_OFFSET);
-        cairo_stroke(GLOBALS->cr_wavepixmap_wavewindow_c_1);
-
-        hashoffset += GLOBALS->hashstep;
-        dx = dx + GLOBALS->hashstep;
-        if ((GLOBALS->pixelsperframe != 200) || (GLOBALS->hashstep != 10.0))
-            iter++; /* fix any roundoff errors */
-        x = dx;
-    }
-}
-
-static void rendertimes(void)
-{
-    int lastx = -1000; /* arbitrary */
-    int x, lenhalf;
-    TimeType tim, rem;
-    char timebuff[32];
-    char prevover = 0;
-    gdouble realx;
-    int s_ctx_iter;
-    int timearray_encountered = 0;
-    static const double dashed1[] = {8.0, 8.0};
-
-    renderblackout();
-
-    tim = GLOBALS->tims.start;
-
-    GLOBALS->tims.end = GLOBALS->tims.start + (((gdouble)GLOBALS->wavewidth) * GLOBALS->nspx);
-
-    /***********/
-    WAVE_STRACE_ITERATOR_FWD(s_ctx_iter)
-    {
-        wave_rgb_t gc;
-
-        if (!s_ctx_iter) {
-            gc = GLOBALS->rgb_gc.gc_grid_wavewindow_c_1;
-            cairo_set_dash(GLOBALS->cr_wavepixmap_wavewindow_c_1, dashed1, 0, 0);
-        } else {
-            gc = GLOBALS->rgb_gc.gc_grid2_wavewindow_c_1;
-            /* gdk_gc_set_line_attributes(gc, 1, GDK_LINE_ON_OFF_DASH, GDK_CAP_BUTT,
-             * GDK_JOIN_BEVEL); */
-            cairo_set_dash(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                           dashed1,
-                           sizeof(dashed1) / sizeof(dashed1[0]),
-                           0);
-        }
-
-        cairo_set_source_rgba(GLOBALS->cr_wavepixmap_wavewindow_c_1, gc.r, gc.g, gc.b, gc.a);
-
-        GLOBALS->strace_ctx = &GLOBALS->strace_windows[GLOBALS->strace_current_window = s_ctx_iter];
-
-        if (GLOBALS->strace_ctx->timearray) {
-            int pos, pos2;
-            TimeType *t, tm;
-            int y = GLOBALS->fontheight + 2;
-            int oldx = -1;
-
-            timearray_encountered = 1;
-
-            pos = bsearch_timechain(GLOBALS->tims.start);
-        top:
-            if ((pos >= 0) && (pos < GLOBALS->strace_ctx->timearray_size)) {
-                t = GLOBALS->strace_ctx->timearray + pos;
-                for (; pos < GLOBALS->strace_ctx->timearray_size; t++, pos++) {
-                    tm = *t;
-                    if (tm >= GLOBALS->tims.start) {
-                        if (tm <= GLOBALS->tims.end) {
-                            x = (tm - GLOBALS->tims.start) * GLOBALS->pxns;
-                            if (oldx == x) {
-                                pos2 = bsearch_timechain(GLOBALS->tims.start +
-                                                         (((gdouble)(x + 1)) * GLOBALS->nspx));
-                                if (pos2 > pos) {
-                                    pos = pos2;
-                                    goto top;
-                                } else
-                                    continue;
-                            }
-                            oldx = x;
-
-                            cairo_move_to(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          x + WAVE_CAIRO_050_OFFSET,
-                                          y + WAVE_CAIRO_050_OFFSET);
-                            cairo_line_to(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          x + WAVE_CAIRO_050_OFFSET,
-                                          GLOBALS->waveheight + WAVE_CAIRO_050_OFFSET);
-                            cairo_stroke(GLOBALS->cr_wavepixmap_wavewindow_c_1);
-                        } else {
-                            break;
-                        }
-                    }
-                }
-            }
-        }
-
-        if (s_ctx_iter) {
-            /* gdk_gc_set_line_attributes(gc, 1, GDK_LINE_SOLID, GDK_CAP_BUTT, GDK_JOIN_BEVEL); */
-        }
-    }
-
-    GLOBALS->strace_ctx = &GLOBALS->strace_windows[GLOBALS->strace_current_window = 0];
-    /***********/
-
-    cairo_set_dash(GLOBALS->cr_wavepixmap_wavewindow_c_1, dashed1, 0, 0);
-
-    if (GLOBALS->ruler_step && !timearray_encountered) {
-        TimeType rhs =
-            (GLOBALS->tims.end > GLOBALS->tims.last) ? GLOBALS->tims.last : GLOBALS->tims.end;
-        TimeType low_x = (GLOBALS->tims.start - GLOBALS->ruler_origin) / GLOBALS->ruler_step;
-        TimeType high_x = (rhs - GLOBALS->ruler_origin) / GLOBALS->ruler_step;
-        TimeType iter_x, tm;
-        int y = GLOBALS->fontheight + 2;
-        int oldx = -1;
-
-        cairo_set_source_rgba(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                              GLOBALS->rgb_gc.gc_grid_wavewindow_c_1.r,
-                              GLOBALS->rgb_gc.gc_grid_wavewindow_c_1.g,
-                              GLOBALS->rgb_gc.gc_grid_wavewindow_c_1.b,
-                              GLOBALS->rgb_gc.gc_grid_wavewindow_c_1.a);
-
-        for (iter_x = low_x; iter_x <= high_x; iter_x++) {
-            tm = GLOBALS->ruler_step * iter_x + GLOBALS->ruler_origin;
-            x = (tm - GLOBALS->tims.start) * GLOBALS->pxns;
-            if (oldx == x) {
-                gdouble xd, offset, pixstep;
-                TimeType newcurr;
-
-                xd = x + 1; /* for pix time calc */
-
-                pixstep = ((gdouble)GLOBALS->nsperframe) / ((gdouble)GLOBALS->pixelsperframe);
-                newcurr = (TimeType)(offset = ((gdouble)GLOBALS->tims.start) + (xd * pixstep));
-
-                if (offset - newcurr > 0.5) /* round to nearest integer ns */
-                {
-                    newcurr++;
-                }
-
-                low_x = (newcurr - GLOBALS->ruler_origin) / GLOBALS->ruler_step;
-                if (low_x <= iter_x)
-                    low_x = (iter_x + 1);
-                iter_x = low_x;
-                tm = GLOBALS->ruler_step * iter_x + GLOBALS->ruler_origin;
-                x = (tm - GLOBALS->tims.start) * GLOBALS->pxns;
-            }
-
-            if (x >= GLOBALS->wavewidth)
-                break;
-            oldx = x;
-
-            cairo_move_to(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                          x + WAVE_CAIRO_050_OFFSET,
-                          y + WAVE_CAIRO_050_OFFSET);
-            cairo_line_to(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                          x + WAVE_CAIRO_050_OFFSET,
-                          GLOBALS->waveheight + WAVE_CAIRO_050_OFFSET);
-            cairo_stroke(GLOBALS->cr_wavepixmap_wavewindow_c_1);
-        }
-    }
-
-    /***********/
-
-    DEBUG(printf("Ruler Start time: " TTFormat ", Finish time: " TTFormat "\n",
-                 GLOBALS->tims.start,
-                 GLOBALS->tims.end));
-
-    x = 0;
-    realx = 0;
-    if (tim) {
-        rem = tim % ((TimeType)GLOBALS->nsperframe);
-        if (rem) {
-            tim = tim - GLOBALS->nsperframe - rem;
-            x = -GLOBALS->pixelsperframe - ((rem * GLOBALS->pixelsperframe) / GLOBALS->nsperframe);
-            realx =
-                -GLOBALS->pixelsperframe - ((rem * GLOBALS->pixelsperframe) / GLOBALS->nsperframe);
-        }
-    }
-
-    for (;;) {
-        renderhash(realx, tim);
-
-        if (tim + GLOBALS->global_time_offset) {
-            if (tim != GLOBALS->min_time) {
-                reformat_time(timebuff,
-                              time_trunc(tim) + GLOBALS->global_time_offset,
-                              GLOBALS->time_dimension);
-            } else {
-                timebuff[0] = 0;
-            }
-        } else {
-            strcpy(timebuff, "0");
-        }
-
-        lenhalf = font_engine_string_measure(GLOBALS->wavefont, timebuff) >> 1;
-
-        if ((x - lenhalf >= lastx) || (GLOBALS->pixelsperframe >= 200)) {
-            XXX_font_engine_draw_string(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                        GLOBALS->wavefont,
-                                        &GLOBALS->rgb_gc.gc_time_wavewindow_c_1,
-                                        x - lenhalf + WAVE_CAIRO_050_OFFSET,
-                                        GLOBALS->wavefont->ascent - 1 + WAVE_CAIRO_050_OFFSET,
-                                        timebuff);
-            lastx = x + lenhalf;
-        }
-
-        tim += GLOBALS->nsperframe;
-        x += GLOBALS->pixelsperframe;
-        realx += GLOBALS->pixelsperframe;
-        if ((prevover) || (tim > GLOBALS->tims.last))
-            break;
-        if (x >= GLOBALS->wavewidth)
-            prevover = 1;
-    }
-}
-
-/***************************************************************************/
-
-static void rendertimebar(void)
-{
-    XXX_gdk_draw_rectangle(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                           GLOBALS->rgb_gc.gc_timeb_wavewindow_c_1,
-                           TRUE,
-                           0,
-                           -1,
-                           GLOBALS->wavewidth,
-                           GLOBALS->fontheight);
-    rendertimes();
-    rendertraces();
-
-    update_dual();
-}
-
-static void gc_save(Trptr t, struct wave_rgbmaster_t *gc_sav)
-{
-    if ((!GLOBALS->black_and_white) && (t->t_color)) {
-        int color = t->t_color;
-
-        color--;
-
-        memcpy(gc_sav, &GLOBALS->rgb_gc, sizeof(struct wave_rgbmaster_t));
-
-        if (color < WAVE_NUM_RAINBOW) {
-            XXX_set_alternate_gcs(GLOBALS->rgb_gc_rainbow[2 * color],
-                                  GLOBALS->rgb_gc_rainbow[2 * color + 1]);
-        }
-    }
-}
-
-static void gc_restore(Trptr t, struct wave_rgbmaster_t *gc_sav)
-{
-    if ((!GLOBALS->black_and_white) && (t->t_color)) {
-        memcpy(&GLOBALS->rgb_gc, gc_sav, sizeof(struct wave_rgbmaster_t));
-    }
-}
-
-static void rendertraces(void)
-{
-    struct wave_rgbmaster_t gc_sav;
-
-    Trptr t = gw_signal_list_get_trace(GW_SIGNAL_LIST(GLOBALS->signalarea), 0);
-    if (t) {
-        Trptr tback = t;
-        hptr h;
-        vptr v;
-        int i = 0, num_traces_displayable;
-        int iback = 0;
-
-        GtkAllocation allocation;
-        gtk_widget_get_allocation(GLOBALS->wavearea, &allocation);
-
-        num_traces_displayable = allocation.height / (GLOBALS->fontheight);
-        num_traces_displayable--; /* for the time trace that is always there */
-
-        /* ensure that transaction traces are visible even if the topmost traces are blanks */
-        while (tback) {
-            if (tback->flags & (TR_BLANK | TR_ANALOG_BLANK_STRETCH)) {
-                tback = GivePrevTrace(tback);
-                iback--;
-            } else if (tback->flags & TR_TTRANSLATED) {
-                if (tback != t) {
-                    t = tback;
-                    i = iback;
-                }
-                break;
-            } else {
-                break;
-            }
-        }
-
-        for (; ((i < num_traces_displayable) && (t)); i++) {
-            if (!(t->flags & (TR_EXCLUDE | TR_BLANK | TR_ANALOG_BLANK_STRETCH))) {
-                GLOBALS->shift_timebase = t->shift;
-                if (!t->vector) {
-                    h = bsearch_node(t->n.nd, GLOBALS->tims.start - t->shift);
-                    DEBUG(printf("Start time: " TTFormat ", Histent time: " TTFormat "\n",
-                                 GLOBALS->tims.start,
-                                 (h->time + GLOBALS->shift_timebase)));
-
-                    if (!t->n.nd->extvals) {
-                        if (i >= 0) {
-                            gc_save(t, &gc_sav);
-                            draw_hptr_trace(t, h, i, 1, 0);
-                            gc_restore(t, &gc_sav);
-                        }
-                    } else {
-                        if (i >= 0) {
-                            gc_save(t, &gc_sav);
-                            draw_hptr_trace_vector(t, h, i);
-                            gc_restore(t, &gc_sav);
-                        }
-                    }
-                } else {
-                    Trptr t_orig, tn;
-                    bvptr bv = t->n.vec;
-
-                    v = bsearch_vector(bv, GLOBALS->tims.start - t->shift);
-                    DEBUG(printf("Vector Trace: %s, %s\n", t->name, bv->bvname));
-                    DEBUG(printf("Start time: " TTFormat ", Vectorent time: " TTFormat "\n",
-                                 GLOBALS->tims.start,
-                                 (v->time + GLOBALS->shift_timebase)));
-                    if (i >= 0) {
-                        gc_save(t, &gc_sav);
-                        draw_vptr_trace(t, v, i);
-                        gc_restore(t, &gc_sav);
-                    }
-
-                    if ((bv->transaction_chain) && (t->flags & TR_TTRANSLATED)) {
-                        t_orig = t;
-                        for (;;) {
-                            tn = GiveNextTrace(t);
-                            bv = bv->transaction_chain;
-
-                            if (bv && tn && (tn->flags & (TR_BLANK | TR_ANALOG_BLANK_STRETCH))) {
-                                i++;
-                                if (i < num_traces_displayable) {
-                                    v = bsearch_vector(bv, GLOBALS->tims.start - t->shift);
-                                    if (i >= 0) {
-                                        gc_save(t, &gc_sav);
-                                        draw_vptr_trace(t_orig, v, i);
-                                        gc_restore(t, &gc_sav);
-                                    }
-                                    t = tn;
-                                    continue;
-                                }
-                            }
-                            break;
-                        }
-                    }
-                }
-            } else {
-                int kill_dodraw_grid = t->flags & TR_ANALOG_BLANK_STRETCH;
-
-                if (kill_dodraw_grid) {
-                    Trptr tn = GiveNextTrace(t);
-                    if (!tn) {
-                        kill_dodraw_grid = 0;
-                    } else if (!(tn->flags & TR_ANALOG_BLANK_STRETCH)) {
-                        kill_dodraw_grid = 0;
-                    }
-                }
-
-                if (i >= 0) {
-                    gc_save(t, &gc_sav);
-                    draw_hptr_trace(NULL, NULL, i, 0, kill_dodraw_grid);
-                    gc_restore(t, &gc_sav);
-                }
-            }
-            t = GiveNextTrace(t);
-            /* bot:		1; */
-        }
-    }
-
-    draw_named_markers();
-    draw_marker_partitions();
-
-    if (GLOBALS->traces.dirty) {
-        char dbuf[32];
-        sprintf(dbuf, "%d", GLOBALS->traces.total);
-        GLOBALS->traces.dirty = 0;
-        gtkwavetcl_setvar(WAVE_TCLCB_TRACES_UPDATED, dbuf, WAVE_TCLCB_TRACES_UPDATED_FLAGS);
-    }
-
-#ifdef GDK_WINDOWING_WAYLAND
-    if (GDK_IS_WAYLAND_DISPLAY(gdk_display_get_default()))
-        gtk_widget_queue_draw(GLOBALS->wavearea);
-#endif
-}
-
-/*
- * draw single traces and use this for rendering the grid lines
- * for "excluded" traces
- */
-static void draw_hptr_trace(Trptr t, hptr h, int which, int dodraw, int kill_grid)
-{
-    TimeType _x0, _x1, newtime;
-    int _y0, _y1, yu, liney, ytext;
-    TimeType tim, h2tim;
-    hptr h2, h3;
-    char hval, h2val, invert;
-    wave_rgb_t c;
-    wave_rgb_t gcx, gcxf;
-    char identifier_str[2];
-    int is_event = t && t->n.nd && (t->n.nd->vartype == ND_VCD_EVENT);
-
-    GLOBALS->tims.start -= GLOBALS->shift_timebase;
-    GLOBALS->tims.end -= GLOBALS->shift_timebase;
-
-    liney = ((which + 2) * GLOBALS->fontheight) - 2;
-    if (((t) && (t->flags & TR_INVERT)) && (!is_event)) {
-        _y0 = ((which + 1) * GLOBALS->fontheight) + 2;
-        _y1 = liney - 2;
-        invert = 1;
-    } else {
-        _y1 = ((which + 1) * GLOBALS->fontheight) + 2;
-        _y0 = liney - 2;
-        invert = 0;
-    }
-    yu = (_y0 + _y1) / 2;
-    ytext = yu - (GLOBALS->wavefont->ascent / 2) + GLOBALS->wavefont->ascent;
-
-    if ((GLOBALS->highlight_wavewindow) && (t) && (t->flags & TR_HIGHLIGHT) &&
-        (!GLOBALS->black_and_white) && (!kill_grid)) {
-        XXX_gdk_draw_rectangle(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                               GLOBALS->rgb_gc.gc_grid_wavewindow_c_1,
-                               TRUE,
-                               0,
-                               liney - GLOBALS->fontheight,
-                               GLOBALS->wavewidth,
-                               GLOBALS->fontheight);
-    } else if ((GLOBALS->display_grid) && (GLOBALS->enable_horiz_grid) && (!kill_grid)) {
-        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                          GLOBALS->rgb_gc.gc_grid_wavewindow_c_1,
-                          (GLOBALS->tims.start < GLOBALS->tims.first)
-                              ? (GLOBALS->tims.first - GLOBALS->tims.start) * GLOBALS->pxns
-                              : 0,
-                          liney,
-                          (GLOBALS->tims.last <= GLOBALS->tims.end)
-                              ? (GLOBALS->tims.last - GLOBALS->tims.start) * GLOBALS->pxns
-                              : GLOBALS->wavewidth - 1,
-                          liney);
-    }
-
-    if ((h) && (GLOBALS->tims.start == h->time))
-        if (h->v.h_val != AN_Z) {
-            switch (h->v.h_val) {
-                case AN_X:
-                    c = GLOBALS->rgb_gc.gc_x_wavewindow_c_1;
-                    break;
-                case AN_U:
-                    c = GLOBALS->rgb_gc.gc_u_wavewindow_c_1;
-                    break;
-                case AN_W:
-                    c = GLOBALS->rgb_gc.gc_w_wavewindow_c_1;
-                    break;
-                case AN_DASH:
-                    c = GLOBALS->rgb_gc.gc_dash_wavewindow_c_1;
-                    break;
-                default:
-                    c = (h->v.h_val == AN_X) ? GLOBALS->rgb_gc.gc_x_wavewindow_c_1
-                                             : GLOBALS->rgb_gc.gc_trans_wavewindow_c_1;
-            }
-            XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1, c, 0, _y0, 0, _y1);
-        }
-
-    if (dodraw && t)
-        for (;;) {
-            if (!h)
-                break;
-            tim = (h->time);
-
-            if ((tim > GLOBALS->tims.end) || (tim > GLOBALS->tims.last))
-                break;
-
-            _x0 = (tim - GLOBALS->tims.start) * GLOBALS->pxns;
-            if (_x0 < -1) {
-                _x0 = -1;
-            } else if (_x0 > GLOBALS->wavewidth) {
-                break;
-            }
-
-            h2 = h->next;
-            if (!h2)
-                break;
-            h2tim = tim = (h2->time);
-            if (tim > GLOBALS->tims.last)
-                tim = GLOBALS->tims.last;
-            else if (tim > GLOBALS->tims.end + 1)
-                tim = GLOBALS->tims.end + 1;
-            _x1 = (tim - GLOBALS->tims.start) * GLOBALS->pxns;
-            if (_x1 < -1) {
-                _x1 = -1;
-            } else if (_x1 > GLOBALS->wavewidth) {
-                _x1 = GLOBALS->wavewidth;
-            }
-
-            if (_x0 != _x1) {
-                if (is_event) {
-                    if (h->time >= GLOBALS->tims.first) {
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          GLOBALS->rgb_gc.gc_w_wavewindow_c_1,
-                                          _x0,
-                                          _y0,
-                                          _x0,
-                                          _y1);
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          GLOBALS->rgb_gc.gc_w_wavewindow_c_1,
-                                          _x0,
-                                          _y1,
-                                          _x0 + 2,
-                                          _y1 + 2);
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          GLOBALS->rgb_gc.gc_w_wavewindow_c_1,
-                                          _x0,
-                                          _y1,
-                                          _x0 - 2,
-                                          _y1 + 2);
-                    }
-                    h = h->next;
-                    continue;
-                }
-
-                hval = h->v.h_val;
-                h2val = h2->v.h_val;
-
-                switch (h2val) {
-                    case AN_X:
-                        c = GLOBALS->rgb_gc.gc_x_wavewindow_c_1;
-                        break;
-                    case AN_U:
-                        c = GLOBALS->rgb_gc.gc_u_wavewindow_c_1;
-                        break;
-                    case AN_W:
-                        c = GLOBALS->rgb_gc.gc_w_wavewindow_c_1;
-                        break;
-                    case AN_DASH:
-                        c = GLOBALS->rgb_gc.gc_dash_wavewindow_c_1;
-                        break;
-                    default:
-                        c = (hval == AN_X) ? GLOBALS->rgb_gc.gc_x_wavewindow_c_1
-                                           : GLOBALS->rgb_gc.gc_trans_wavewindow_c_1;
-                }
-
-                switch (hval) {
-                    case AN_0: /* 0 */
-                    case AN_L: /* L */
-                        if (GLOBALS->fill_waveform && invert) {
-                            switch (hval) {
-                                case AN_0:
-                                    gcxf = GLOBALS->rgb_gc.gc_1fill_wavewindow_c_1;
-                                    break;
-                                case AN_L:
-                                    gcxf = GLOBALS->rgb_gc.gc_highfill_wavewindow_c_1;
-                                    break;
-                            }
-                            XXX_gdk_draw_rectangle(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                                   gcxf,
-                                                   TRUE,
-                                                   _x0 + 1,
-                                                   _y0,
-                                                   _x1 - _x0,
-                                                   _y1 - _y0 + 1);
-                        }
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          (hval == AN_0) ? GLOBALS->rgb_gc.gc_0_wavewindow_c_1
-                                                         : GLOBALS->rgb_gc.gc_low_wavewindow_c_1,
-                                          _x0,
-                                          _y0,
-                                          _x1,
-                                          _y0);
-
-                        if (h2tim <= GLOBALS->tims.end)
-                            switch (h2val) {
-                                case AN_0:
-                                case AN_L:
-                                    break;
-
-                                case AN_Z:
-                                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                                      c,
-                                                      _x1,
-                                                      _y0,
-                                                      _x1,
-                                                      yu);
-                                    break;
-                                default:
-                                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                                      c,
-                                                      _x1,
-                                                      _y0,
-                                                      _x1,
-                                                      _y1);
-                                    break;
-                            }
-                        break;
-
-                    case AN_X: /* X */
-                    case AN_W: /* W */
-                    case AN_U: /* U */
-                    case AN_DASH: /* - */
-
-                        identifier_str[1] = 0;
-                        switch (hval) {
-                            case AN_X:
-                                c = gcx = GLOBALS->rgb_gc.gc_x_wavewindow_c_1;
-                                gcxf = GLOBALS->rgb_gc.gc_xfill_wavewindow_c_1;
-                                identifier_str[0] = 0;
-                                break;
-                            case AN_W:
-                                c = gcx = GLOBALS->rgb_gc.gc_w_wavewindow_c_1;
-                                gcxf = GLOBALS->rgb_gc.gc_wfill_wavewindow_c_1;
-                                identifier_str[0] = 'W';
-                                break;
-                            case AN_U:
-                                c = gcx = GLOBALS->rgb_gc.gc_u_wavewindow_c_1;
-                                gcxf = GLOBALS->rgb_gc.gc_ufill_wavewindow_c_1;
-                                identifier_str[0] = 'U';
-                                break;
-                            default:
-                                c = gcx = GLOBALS->rgb_gc.gc_dash_wavewindow_c_1;
-                                gcxf = GLOBALS->rgb_gc.gc_dashfill_wavewindow_c_1;
-                                identifier_str[0] = '-';
-                                break;
-                        }
-
-                        if (invert) {
-                            XXX_gdk_draw_rectangle(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                                   gcx,
-                                                   TRUE,
-                                                   _x0 + 1,
-                                                   _y0,
-                                                   _x1 - _x0,
-                                                   _y1 - _y0 + 1);
-                        } else {
-                            XXX_gdk_draw_rectangle(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                                   gcxf,
-                                                   TRUE,
-                                                   _x0 + 1,
-                                                   _y1,
-                                                   _x1 - _x0,
-                                                   _y0 - _y1 + 1);
-                        }
-
-                        if (identifier_str[0]) {
-                            int _x0_new = (_x0 >= 0) ? _x0 : 0;
-                            int width;
-
-                            if ((width = _x1 - _x0_new) > GLOBALS->vector_padding) {
-                                if ((_x1 >= GLOBALS->wavewidth) ||
-                                    (font_engine_string_measure(GLOBALS->wavefont, identifier_str) +
-                                         GLOBALS->vector_padding <=
-                                     width)) {
-                                    XXX_font_engine_draw_string(
-                                        GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                        GLOBALS->wavefont,
-                                        &GLOBALS->rgb_gc.gc_value_wavewindow_c_1,
-                                        _x0 + 2 + WAVE_CAIRO_050_OFFSET,
-                                        ytext + WAVE_CAIRO_050_OFFSET,
-                                        identifier_str);
-                                }
-                            }
-                        }
-
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          gcx,
-                                          _x0,
-                                          _y0,
-                                          _x1,
-                                          _y0);
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          gcx,
-                                          _x0,
-                                          _y1,
-                                          _x1,
-                                          _y1);
-                        if (h2tim <= GLOBALS->tims.end)
-                            XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                              c,
-                                              _x1,
-                                              _y0,
-                                              _x1,
-                                              _y1);
-                        break;
-
-                    case AN_Z: /* Z */
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          GLOBALS->rgb_gc.gc_mid_wavewindow_c_1,
-                                          _x0,
-                                          yu,
-                                          _x1,
-                                          yu);
-                        if (h2tim <= GLOBALS->tims.end)
-                            switch (h2val) {
-                                case AN_0:
-                                case AN_L:
-                                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                                      c,
-                                                      _x1,
-                                                      yu,
-                                                      _x1,
-                                                      _y0);
-                                    break;
-                                case AN_1:
-                                case AN_H:
-                                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                                      c,
-                                                      _x1,
-                                                      yu,
-                                                      _x1,
-                                                      _y1);
-                                    break;
-                                default:
-                                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                                      c,
-                                                      _x1,
-                                                      _y0,
-                                                      _x1,
-                                                      _y1);
-                                    break;
-                            }
-                        break;
-
-                    case AN_1: /* 1 */
-                    case AN_H: /* 1 */
-                        if (GLOBALS->fill_waveform && !invert) {
-                            switch (hval) {
-                                case AN_1:
-                                    gcxf = GLOBALS->rgb_gc.gc_1fill_wavewindow_c_1;
-                                    break;
-                                case AN_H:
-                                    gcxf = GLOBALS->rgb_gc.gc_highfill_wavewindow_c_1;
-                                    break;
-                            }
-                            XXX_gdk_draw_rectangle(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                                   gcxf,
-                                                   TRUE,
-                                                   _x0 + 1,
-                                                   _y1,
-                                                   _x1 - _x0,
-                                                   _y0 - _y1 + 1);
-                        }
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          (hval == AN_1) ? GLOBALS->rgb_gc.gc_1_wavewindow_c_1
-                                                         : GLOBALS->rgb_gc.gc_high_wavewindow_c_1,
-                                          _x0,
-                                          _y1,
-                                          _x1,
-                                          _y1);
-                        if (h2tim <= GLOBALS->tims.end)
-                            switch (h2val) {
-                                case AN_1:
-                                case AN_H:
-                                    break;
-
-                                case AN_0:
-                                case AN_L:
-                                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                                      c,
-                                                      _x1,
-                                                      _y1,
-                                                      _x1,
-                                                      _y0);
-                                    break;
-                                case AN_Z:
-                                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                                      c,
-                                                      _x1,
-                                                      _y1,
-                                                      _x1,
-                                                      yu);
-                                    break;
-                                default:
-                                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                                      c,
-                                                      _x1,
-                                                      _y0,
-                                                      _x1,
-                                                      _y1);
-                                    break;
-                            }
-                        break;
-
-                    default:
-                        break;
-                }
-            } else {
-                if (!is_event) {
-                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                      GLOBALS->rgb_gc.gc_trans_wavewindow_c_1,
-                                      _x1,
-                                      _y0,
-                                      _x1,
-                                      _y1);
-                } else {
-                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                      GLOBALS->rgb_gc.gc_w_wavewindow_c_1,
-                                      _x1,
-                                      _y0,
-                                      _x1,
-                                      _y1);
-                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                      GLOBALS->rgb_gc.gc_w_wavewindow_c_1,
-                                      _x0,
-                                      _y1,
-                                      _x0 + 2,
-                                      _y1 + 2);
-                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                      GLOBALS->rgb_gc.gc_w_wavewindow_c_1,
-                                      _x0,
-                                      _y1,
-                                      _x0 - 2,
-                                      _y1 + 2);
-                }
-                newtime = (((gdouble)(_x1 + WAVE_OPT_SKIP)) * GLOBALS->nspx) +
-                          GLOBALS->tims.start /*+GLOBALS->shift_timebase*/; /* skip to next pixel */
-                h3 = bsearch_node(t->n.nd, newtime);
-                if (h3->time > h->time) {
-                    h = h3;
-                    continue;
-                }
-            }
-
-            if ((h->flags & HIST_GLITCH) && (GLOBALS->vcd_preserve_glitches)) {
-                XXX_gdk_draw_rectangle(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                       GLOBALS->rgb_gc.gc_mid_wavewindow_c_1,
-                                       TRUE,
-                                       _x1 - 1,
-                                       yu - 1,
-                                       3,
-                                       3);
-            }
-
-            h = h->next;
-        }
-
-    GLOBALS->tims.start += GLOBALS->shift_timebase;
-    GLOBALS->tims.end += GLOBALS->shift_timebase;
-}
-
-/********************************************************************************************************/
-
-static void draw_hptr_trace_vector_analog(Trptr t, hptr h, int which, int num_extension)
-{
-    TimeType _x0, _x1, newtime;
-    int _y0, _y1, yu, liney, yt0, yt1;
-    TimeType tim, h2tim;
-    hptr h2, h3;
-    int endcnt = 0;
-    int type;
-    /* int lasttype=-1; */ /* scan-build */
-    wave_rgb_t c, ci;
-    wave_rgb_t cnan = GLOBALS->rgb_gc.gc_u_wavewindow_c_1;
-    wave_rgb_t cinf = GLOBALS->rgb_gc.gc_w_wavewindow_c_1;
-    wave_rgb_t cfixed;
-    double mynan = strtod("NaN", NULL);
-    double tmin = mynan, tmax = mynan, tv, tv2;
-    gint rmargin;
-    int is_nan = 0, is_nan2 = 0, is_inf = 0, is_inf2 = 0;
-    int any_infs = 0, any_infp = 0, any_infm = 0;
-    int skipcnt = 0;
-
-    ci = GLOBALS->rgb_gc.gc_baseline_wavewindow_c_1;
-
-    liney = ((which + 2 + num_extension) * GLOBALS->fontheight) - 2;
-    _y1 = ((which + 1) * GLOBALS->fontheight) + 2;
-    _y0 = liney - 2;
-    yu = (_y0 + _y1) / 2;
-
-    if (t->flags & TR_ANALOG_FULLSCALE) /* otherwise use dynamic */
-    {
-        if ((!t->minmax_valid) || (t->d_num_ext != num_extension)) {
-            h3 = &t->n.nd->head;
-            for (;;) {
-                if (!h3)
-                    break;
-
-                if ((h3->time >= GLOBALS->tims.first) && (h3->time <= GLOBALS->tims.last)) {
-                    tv = mynan;
-                    if (h3->flags & HIST_REAL) {
-#ifdef WAVE_HAS_H_DOUBLE
-                        if (!(h3->flags & HIST_STRING))
-                            tv = h3->v.h_double;
-#else
-                        if (!(h3->flags & HIST_STRING) && h3->v.h_vector)
-                            tv = *(double *)h3->v.h_vector;
-#endif
-                    } else {
-                        if (h3->time <= GLOBALS->tims.last)
-                            tv = convert_real_vec(t, h3->v.h_vector);
-                    }
-
-                    if (!isnan(tv) && !isinf(tv)) {
-                        if (isnan(tmin) || tv < tmin)
-                            tmin = tv;
-                        if (isnan(tmax) || tv > tmax)
-                            tmax = tv;
-                    } else if (isinf(tv)) {
-                        any_infs = 1;
-
-                        if (tv > 0) {
-                            any_infp = 1;
-                        } else {
-                            any_infm = 1;
-                        }
-                    }
-                }
-                h3 = h3->next;
-            }
-
-            if (isnan(tmin) || isnan(tmax)) {
-                tmin = tmax = 0;
-            }
-
-            if (any_infs) {
-                double tdelta = (tmax - tmin) * WAVE_INF_SCALING;
-
-                if (any_infp)
-                    tmax = tmax + tdelta;
-                if (any_infm)
-                    tmin = tmin - tdelta;
-            }
-
-            if ((tmax - tmin) < 1e-20) {
-                tmax = 1;
-                tmin -= 0.5 * (_y1 - _y0);
-            } else {
-                tmax = (_y1 - _y0) / (tmax - tmin);
-            }
-
-            t->minmax_valid = 1;
-            t->d_minval = tmin;
-            t->d_maxval = tmax;
-            t->d_num_ext = num_extension;
-        } else {
-            tmin = t->d_minval;
-            tmax = t->d_maxval;
-        }
-    } else {
-        h3 = h;
-        for (;;) {
-            if (!h3)
-                break;
-            tim = (h3->time);
-            if (tim > GLOBALS->tims.end) {
-                endcnt++;
-                if (endcnt == 2)
-                    break;
-            }
-            if (tim > GLOBALS->tims.last)
-                break;
-
-            _x0 = (tim - GLOBALS->tims.start) * GLOBALS->pxns;
-            if ((_x0 > GLOBALS->wavewidth) && (endcnt == 2)) {
-                break;
-            }
-
-            tv = mynan;
-            if (h3->flags & HIST_REAL) {
-#ifdef WAVE_HAS_H_DOUBLE
-                if (!(h3->flags & HIST_STRING))
-                    tv = h3->v.h_double;
-#else
-                if (!(h3->flags & HIST_STRING) && h3->v.h_vector)
-                    tv = *(double *)h3->v.h_vector;
-#endif
-            } else {
-                if (h3->time <= GLOBALS->tims.last)
-                    tv = convert_real_vec(t, h3->v.h_vector);
-            }
-
-            if (!isnan(tv) && !isinf(tv)) {
-                if (isnan(tmin) || tv < tmin)
-                    tmin = tv;
-                if (isnan(tmax) || tv > tmax)
-                    tmax = tv;
-            } else if (isinf(tv)) {
-                any_infs = 1;
-                if (tv > 0) {
-                    any_infp = 1;
-                } else {
-                    any_infm = 1;
-                }
-            }
-
-            h3 = h3->next;
-        }
-
-        if (isnan(tmin) || isnan(tmax))
-            tmin = tmax = 0;
-
-        if (any_infs) {
-            double tdelta = (tmax - tmin) * WAVE_INF_SCALING;
-
-            if (any_infp)
-                tmax = tmax + tdelta;
-            if (any_infm)
-                tmin = tmin - tdelta;
-        }
-
-        if ((tmax - tmin) < 1e-20) {
-            tmax = 1;
-            tmin -= 0.5 * (_y1 - _y0);
-        } else {
-            tmax = (_y1 - _y0) / (tmax - tmin);
-        }
-    }
-
-    if (GLOBALS->tims.last - GLOBALS->tims.start < GLOBALS->wavewidth) {
-        rmargin = (GLOBALS->tims.last - GLOBALS->tims.start) * GLOBALS->pxns;
-    } else {
-        rmargin = GLOBALS->wavewidth;
-    }
-
-    /* now do the actual drawing */
-    h3 = NULL;
-    for (;;) {
-        if (!h)
-            break;
-        tim = (h->time);
-        if ((tim > GLOBALS->tims.end) || (tim > GLOBALS->tims.last))
-            break;
-
-        _x0 = (tim - GLOBALS->tims.start) * GLOBALS->pxns;
-
-        /*
-        if(_x0<-1)
-                {
-                _x0=-1;
-                }
-                else
-        if(_x0>GLOBALS->wavewidth)
-                {
-                break;
-                }
-        */
-
-        h2 = h->next;
-        if (!h2)
-            break;
-        h2tim = tim = (h2->time);
-        if (tim > GLOBALS->tims.last)
-            tim = GLOBALS->tims.last;
-        /*	else if(tim>GLOBALS->tims.end+1) tim=GLOBALS->tims.end+1; */
-        _x1 = (tim - GLOBALS->tims.start) * GLOBALS->pxns;
-
-        /*
-        if(_x1<-1)
-                {
-                _x1=-1;
-                }
-                else
-        if(_x1>GLOBALS->wavewidth)
-                {
-                _x1=GLOBALS->wavewidth;
-                }
-        */
-
-        /* draw trans */
-        type = (!(h->flags & (HIST_REAL | HIST_STRING))) ? vtype(t, h->v.h_vector) : AN_COUNT;
-        tv = tv2 = mynan;
-
-        if (h->flags & HIST_REAL) {
-#ifdef WAVE_HAS_H_DOUBLE
-            if (!(h->flags & HIST_STRING))
-                tv = h->v.h_double;
-#else
-            if (!(h->flags & HIST_STRING) && h->v.h_vector)
-                tv = *(double *)h->v.h_vector;
-#endif
-        } else {
-            if (h->time <= GLOBALS->tims.last)
-                tv = convert_real_vec(t, h->v.h_vector);
-        }
-
-        if (h2->flags & HIST_REAL) {
-#ifdef WAVE_HAS_H_DOUBLE
-            if (!(h2->flags & HIST_STRING))
-                tv2 = h2->v.h_double;
-#else
-            if (!(h2->flags & HIST_STRING) && h2->v.h_vector)
-                tv2 = *(double *)h2->v.h_vector;
-#endif
-        } else {
-            if (h2->time <= GLOBALS->tims.last)
-                tv2 = convert_real_vec(t, h2->v.h_vector);
-        }
-
-        if ((is_inf = isinf(tv))) {
-            if (tv < 0) {
-                yt0 = _y0;
-            } else {
-                yt0 = _y1;
-            }
-        } else if ((is_nan = isnan(tv))) {
-            yt0 = yu;
-        } else {
-            yt0 = _y0 + (tv - tmin) * tmax;
-        }
-
-        if ((is_inf2 = isinf(tv2))) {
-            if (tv2 < 0) {
-                yt1 = _y0;
-            } else {
-                yt1 = _y1;
-            }
-        } else if ((is_nan2 = isnan(tv2))) {
-            yt1 = yu;
-        } else {
-            yt1 = _y0 + (tv2 - tmin) * tmax;
-        }
-
-        if ((_x0 != _x1) ||
-            (skipcnt < GLOBALS->analog_redraw_skip_count)) /* lower number = better performance */
-        {
-            if (_x0 == _x1) {
-                skipcnt++;
-            } else {
-                skipcnt = 0;
-            }
-
-            if (type != AN_X) {
-                c = GLOBALS->rgb_gc.gc_vbox_wavewindow_c_1;
-            } else {
-                c = GLOBALS->rgb_gc.gc_x_wavewindow_c_1;
-            }
-
-            if (h->next) {
-                if (h->next->time > GLOBALS->max_time) {
-                    yt1 = yt0;
-                }
-            }
-
-            cfixed = is_inf ? cinf : c;
-            if ((is_nan2) && (h2tim > GLOBALS->max_time))
-                is_nan2 = 0;
-
-            /* clamp to top/bottom because of integer rounding errors */
-
-            if (yt0 < _y1)
-                yt0 = _y1;
-            else if (yt0 > _y0)
-                yt0 = _y0;
-
-            if (yt1 < _y1)
-                yt1 = _y1;
-            else if (yt1 > _y0)
-                yt1 = _y0;
-
-            /* clipping... */
-            {
-                int coords[4];
-                int rect[4];
-
-                if (_x0 < INT_MIN) {
-                    coords[0] = INT_MIN;
-                } else if (_x0 > INT_MAX) {
-                    coords[0] = INT_MAX;
-                } else {
-                    coords[0] = _x0;
-                }
-
-                if (_x1 < INT_MIN) {
-                    coords[2] = INT_MIN;
-                } else if (_x1 > INT_MAX) {
-                    coords[2] = INT_MAX;
-                } else {
-                    coords[2] = _x1;
-                }
-
-                coords[1] = yt0;
-                coords[3] = yt1;
-
-                rect[0] = -10;
-                rect[1] = _y1;
-                rect[2] = GLOBALS->wavewidth + 10;
-                rect[3] = _y0;
-
-                if ((t->flags & (TR_ANALOG_INTERPOLATED | TR_ANALOG_STEP)) != TR_ANALOG_STEP) {
-                    wave_lineclip(coords, rect);
-                } else {
-                    if (coords[0] < rect[0])
-                        coords[0] = rect[0];
-                    if (coords[2] < rect[0])
-                        coords[2] = rect[0];
-
-                    if (coords[0] > rect[2])
-                        coords[0] = rect[2];
-                    if (coords[2] > rect[2])
-                        coords[2] = rect[2];
-
-                    if (coords[1] < rect[1])
-                        coords[1] = rect[1];
-                    if (coords[3] < rect[1])
-                        coords[3] = rect[1];
-
-                    if (coords[1] > rect[3])
-                        coords[1] = rect[3];
-                    if (coords[3] > rect[3])
-                        coords[3] = rect[3];
-                }
-
-                _x0 = coords[0];
-                yt0 = coords[1];
-                _x1 = coords[2];
-                yt1 = coords[3];
-            }
-            /* ...clipping */
-
-            if (is_nan || is_nan2) {
-                if (is_nan) {
-                    XXX_gdk_draw_rectangle(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                           cnan,
-                                           TRUE,
-                                           _x0,
-                                           _y1,
-                                           _x1 - _x0,
-                                           _y0 - _y1);
-
-                    if ((t->flags & (TR_ANALOG_INTERPOLATED | TR_ANALOG_STEP)) ==
-                        (TR_ANALOG_INTERPOLATED | TR_ANALOG_STEP)) {
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          ci,
-                                          _x1 - 1,
-                                          yt1,
-                                          _x1 + 1,
-                                          yt1);
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          ci,
-                                          _x1,
-                                          yt1 - 1,
-                                          _x1,
-                                          yt1 + 1);
-
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          ci,
-                                          _x0 - 1,
-                                          _y0,
-                                          _x0 + 1,
-                                          _y0);
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          ci,
-                                          _x0,
-                                          _y0 - 1,
-                                          _x0,
-                                          _y0 + 1);
-
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          ci,
-                                          _x0 - 1,
-                                          _y1,
-                                          _x0 + 1,
-                                          _y1);
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          ci,
-                                          _x0,
-                                          _y1 - 1,
-                                          _x0,
-                                          _y1 + 1);
-                    }
-                }
-                if (is_nan2) {
-                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                      cfixed,
-                                      _x0,
-                                      yt0,
-                                      _x1,
-                                      yt0);
-
-                    if ((t->flags & (TR_ANALOG_INTERPOLATED | TR_ANALOG_STEP)) ==
-                        (TR_ANALOG_INTERPOLATED | TR_ANALOG_STEP)) {
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          cnan,
-                                          _x1,
-                                          yt1,
-                                          _x1,
-                                          yt0);
-
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          ci,
-                                          _x1 - 1,
-                                          _y0,
-                                          _x1 + 1,
-                                          _y0);
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          ci,
-                                          _x1,
-                                          _y0 - 1,
-                                          _x1,
-                                          _y0 + 1);
-
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          ci,
-                                          _x1 - 1,
-                                          _y1,
-                                          _x1 + 1,
-                                          _y1);
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          ci,
-                                          _x1,
-                                          _y1 - 1,
-                                          _x1,
-                                          _y1 + 1);
-                    }
-                }
-            } else if ((t->flags & TR_ANALOG_INTERPOLATED) && !is_inf && !is_inf2) {
-                if (t->flags & TR_ANALOG_STEP) {
-                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                      ci,
-                                      _x0 - 1,
-                                      yt0,
-                                      _x0 + 1,
-                                      yt0);
-                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                      ci,
-                                      _x0,
-                                      yt0 - 1,
-                                      _x0,
-                                      yt0 + 1);
-                }
-
-                if (rmargin != GLOBALS->wavewidth) /* the window is clipped in postscript */
-                {
-                    if ((yt0 == yt1) && ((_x0 > _x1) || (_x0 < 0))) {
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          cfixed,
-                                          0,
-                                          yt0,
-                                          _x1,
-                                          yt1);
-                    } else {
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          cfixed,
-                                          _x0,
-                                          yt0,
-                                          _x1,
-                                          yt1);
-                    }
-                } else {
-                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                      cfixed,
-                                      _x0,
-                                      yt0,
-                                      _x1,
-                                      yt1);
-                }
-            } else
-            /* if(t->flags & TR_ANALOG_STEP) */
-            {
-                XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                  cfixed,
-                                  _x0,
-                                  yt0,
-                                  _x1,
-                                  yt0);
-
-                if (is_inf2)
-                    cfixed = cinf;
-                XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                  cfixed,
-                                  _x1,
-                                  yt0,
-                                  _x1,
-                                  yt1);
-
-                if ((t->flags & (TR_ANALOG_INTERPOLATED | TR_ANALOG_STEP)) ==
-                    (TR_ANALOG_INTERPOLATED | TR_ANALOG_STEP)) {
-                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                      ci,
-                                      _x0 - 1,
-                                      yt0,
-                                      _x0 + 1,
-                                      yt0);
-                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                      ci,
-                                      _x0,
-                                      yt0 - 1,
-                                      _x0,
-                                      yt0 + 1);
-                }
-            }
-        } else {
-            newtime = (((gdouble)(_x1 + WAVE_OPT_SKIP)) * GLOBALS->nspx) +
-                      GLOBALS->tims.start /*+GLOBALS->shift_timebase*/; /* skip to next pixel */
-            h3 = bsearch_node(t->n.nd, newtime);
-            if (h3->time > h->time) {
-                h = h3;
-                /* lasttype=type; */ /* scan-build */
-                continue;
-            }
-        }
-
-        h = h->next;
-        /* lasttype=type; */ /* scan-build */
-    }
-}
-
-/*
- * draw hptr vectors (integer+real)
- */
-static void draw_hptr_trace_vector(Trptr t, hptr h, int which)
-{
-    TimeType _x0, _x1, newtime, width;
-    int _y0, _y1, yu, liney, ytext;
-    TimeType tim /* , h2tim */; /* scan-build */
-    hptr h2, h3;
-    char *ascii = NULL;
-    int type;
-    int lasttype = -1;
-    wave_rgb_t c;
-
-    GLOBALS->tims.start -= GLOBALS->shift_timebase;
-    GLOBALS->tims.end -= GLOBALS->shift_timebase;
-
-    liney = ((which + 2) * GLOBALS->fontheight) - 2;
-    _y1 = ((which + 1) * GLOBALS->fontheight) + 2;
-    _y0 = liney - 2;
-    yu = (_y0 + _y1) / 2;
-    ytext = yu - (GLOBALS->wavefont->ascent / 2) + GLOBALS->wavefont->ascent;
-
-    if ((GLOBALS->highlight_wavewindow) && (t) && (t->flags & TR_HIGHLIGHT) &&
-        (!GLOBALS->black_and_white)) {
-        Trptr tn = GiveNextTrace(t);
-        if ((t->flags & TR_ANALOGMASK) && (tn) && (tn->flags & TR_ANALOG_BLANK_STRETCH)) {
-            XXX_gdk_draw_rectangle(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                   GLOBALS->rgb_gc.gc_grid_wavewindow_c_1,
-                                   TRUE,
-                                   0,
-                                   liney - GLOBALS->fontheight,
-                                   GLOBALS->wavewidth,
-                                   GLOBALS->fontheight);
-        } else {
-            XXX_gdk_draw_rectangle(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                   GLOBALS->rgb_gc.gc_grid_wavewindow_c_1,
-                                   TRUE,
-                                   0,
-                                   liney - GLOBALS->fontheight,
-                                   GLOBALS->wavewidth,
-                                   GLOBALS->fontheight);
-        }
-    } else if ((GLOBALS->display_grid) && (GLOBALS->enable_horiz_grid)) {
-        Trptr tn = GiveNextTrace(t);
-        if ((t->flags & TR_ANALOGMASK) && (tn) && (tn->flags & TR_ANALOG_BLANK_STRETCH)) {
-        } else {
-            XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                              GLOBALS->rgb_gc.gc_grid_wavewindow_c_1,
-                              (GLOBALS->tims.start < GLOBALS->tims.first)
-                                  ? (GLOBALS->tims.first - GLOBALS->tims.start) * GLOBALS->pxns
-                                  : 0,
-                              liney,
-                              (GLOBALS->tims.last <= GLOBALS->tims.end)
-                                  ? (GLOBALS->tims.last - GLOBALS->tims.start) * GLOBALS->pxns
-                                  : GLOBALS->wavewidth - 1,
-                              liney);
-        }
-    }
-
-    if ((t->flags & TR_ANALOGMASK) && (!(h->flags & HIST_STRING) || !(h->flags & HIST_REAL))) {
-        Trptr te = GiveNextTrace(t);
-        int ext = 0;
-
-        while (te) {
-            if (te->flags & TR_ANALOG_BLANK_STRETCH) {
-                ext++;
-                te = GiveNextTrace(te);
-            } else {
-                break;
-            }
-        }
-
-        if ((ext) && (GLOBALS->highlight_wavewindow) && (t) && (t->flags & TR_HIGHLIGHT) &&
-            (!GLOBALS->black_and_white)) {
-            XXX_gdk_draw_rectangle(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                   GLOBALS->rgb_gc.gc_grid_wavewindow_c_1,
-                                   TRUE,
-                                   0,
-                                   liney,
-                                   GLOBALS->wavewidth,
-                                   GLOBALS->fontheight * ext);
-        }
-
-        draw_hptr_trace_vector_analog(t, h, which, ext);
-        GLOBALS->tims.start += GLOBALS->shift_timebase;
-        GLOBALS->tims.end += GLOBALS->shift_timebase;
-        return;
-    }
-
-    GLOBALS->color_active_in_filter = 1;
-
-    for (;;) {
-        if (!h)
-            break;
-        tim = (h->time);
-        if ((tim > GLOBALS->tims.end) || (tim > GLOBALS->tims.last))
-            break;
-
-        _x0 = (tim - GLOBALS->tims.start) * GLOBALS->pxns;
-        if (_x0 < -1) {
-            _x0 = -1;
-        } else if (_x0 > GLOBALS->wavewidth) {
-            break;
-        }
-
-        h2 = h->next;
-        if (!h2)
-            break;
-        /* h2tim= */ tim = (h2->time); /* scan-build */
-        if (tim > GLOBALS->tims.last)
-            tim = GLOBALS->tims.last;
-        else if (tim > GLOBALS->tims.end + 1)
-            tim = GLOBALS->tims.end + 1;
-        _x1 = (tim - GLOBALS->tims.start) * GLOBALS->pxns;
-        if (_x1 < -1) {
-            _x1 = -1;
-        } else if (_x1 > GLOBALS->wavewidth) {
-            _x1 = GLOBALS->wavewidth;
-        }
-
-        /* draw trans */
-        if (!(h->flags & (HIST_REAL | HIST_STRING))) {
-            type = vtype(t, h->v.h_vector);
-        } else {
-            /* s\000 ID is special "z" case */
-            type = AN_COUNT;
-
-            if (h->flags & HIST_STRING) {
-                if (h->v.h_vector) {
-                    if (!h->v.h_vector[0]) {
-                        type = AN_Z;
-                    } else {
-                        if (!strcmp(h->v.h_vector, "UNDEF")) {
-                            type = AN_X;
-                        }
-                    }
-                } else {
-                    type = AN_X;
-                }
-            }
-        }
-        /* type = (!(h->flags&(HIST_REAL|HIST_STRING))) ? vtype(t,h->v.h_vector) : AN_COUNT; */
-        if (_x0 > -1) {
-            wave_rgb_t gltype, gtype;
-
-            switch (lasttype) {
-                case AN_X:
-                    gltype = GLOBALS->rgb_gc.gc_x_wavewindow_c_1;
-                    break;
-                case AN_U:
-                    gltype = GLOBALS->rgb_gc.gc_u_wavewindow_c_1;
-                    break;
-                default:
-                    gltype = GLOBALS->rgb_gc.gc_vtrans_wavewindow_c_1;
-                    break;
-            }
-            switch (type) {
-                case AN_X:
-                    gtype = GLOBALS->rgb_gc.gc_x_wavewindow_c_1;
-                    break;
-                case AN_U:
-                    gtype = GLOBALS->rgb_gc.gc_u_wavewindow_c_1;
-                    break;
-                default:
-                    gtype = GLOBALS->rgb_gc.gc_vtrans_wavewindow_c_1;
-                    break;
-            }
-
-            if (GLOBALS->use_roundcaps) {
-                if (type == AN_Z) {
-                    if (lasttype != -1) {
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          gltype,
-                                          _x0 - 1,
-                                          _y0,
-                                          _x0,
-                                          yu);
-                        if (lasttype != AN_0)
-                            XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                              gltype,
-                                              _x0,
-                                              yu,
-                                              _x0 - 1,
-                                              _y1);
-                    }
-                } else if (lasttype == AN_Z) {
-                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                      gtype,
-                                      _x0 + 1,
-                                      _y0,
-                                      _x0,
-                                      yu);
-                    if (type != AN_0)
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          gtype,
-                                          _x0,
-                                          yu,
-                                          _x0 + 1,
-                                          _y1);
-                } else {
-                    if (lasttype != type) {
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          gltype,
-                                          _x0 - 1,
-                                          _y0,
-                                          _x0,
-                                          yu);
-                        if (lasttype != AN_0)
-                            XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                              gltype,
-                                              _x0,
-                                              yu,
-                                              _x0 - 1,
-                                              _y1);
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          gtype,
-                                          _x0 + 1,
-                                          _y0,
-                                          _x0,
-                                          yu);
-                        if (type != AN_0)
-                            XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                              gtype,
-                                              _x0,
-                                              yu,
-                                              _x0 + 1,
-                                              _y1);
-                    } else {
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          gtype,
-                                          _x0 - 2,
-                                          _y0,
-                                          _x0 + 2,
-                                          _y1);
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          gtype,
-                                          _x0 + 2,
-                                          _y0,
-                                          _x0 - 2,
-                                          _y1);
-                    }
-                }
-            } else {
-                XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1, gtype, _x0, _y0, _x0, _y1);
-            }
-        }
-
-        if (_x0 != _x1) {
-            if (type == AN_Z) {
-                if (GLOBALS->use_roundcaps) {
-                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                      GLOBALS->rgb_gc.gc_mid_wavewindow_c_1,
-                                      _x0 + 1,
-                                      yu,
-                                      _x1 - 1,
-                                      yu);
-                } else {
-                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                      GLOBALS->rgb_gc.gc_mid_wavewindow_c_1,
-                                      _x0,
-                                      yu,
-                                      _x1,
-                                      yu);
-                }
-            } else {
-                if ((type != AN_X) && (type != AN_U)) {
-                    c = GLOBALS->rgb_gc.gc_vbox_wavewindow_c_1;
-                } else {
-                    c = GLOBALS->rgb_gc.gc_x_wavewindow_c_1;
-                }
-
-                if (GLOBALS->use_roundcaps) {
-                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                      c,
-                                      _x0 + 2,
-                                      _y0,
-                                      _x1 - 2,
-                                      _y0);
-                    if (type != AN_0)
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          c,
-                                          _x0 + 2,
-                                          _y1,
-                                          _x1 - 2,
-                                          _y1);
-                    if (type == AN_1)
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          c,
-                                          _x0 + 2,
-                                          _y1 + 1,
-                                          _x1 - 2,
-                                          _y1 + 1);
-                } else {
-                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1, c, _x0, _y0, _x1, _y0);
-                    if (type != AN_0)
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          c,
-                                          _x0,
-                                          _y1,
-                                          _x1,
-                                          _y1);
-                    if (type == AN_1)
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          c,
-                                          _x0,
-                                          _y1 + 1,
-                                          _x1,
-                                          _y1 + 1);
-                }
-
-                if (_x0 < 0)
-                    _x0 = 0; /* fixup left margin */
-
-                if ((width = _x1 - _x0) > GLOBALS->vector_padding) {
-                    char *ascii2;
-
-                    if (h->flags & HIST_REAL) {
-                        if (!(h->flags & HIST_STRING)) {
-#ifdef WAVE_HAS_H_DOUBLE
-                            ascii = convert_ascii_real(t, &h->v.h_double);
-#else
-                            ascii = convert_ascii_real(t, (double *)h->v.h_vector);
-#endif
-                        } else {
-                            ascii = convert_ascii_string((char *)h->v.h_vector);
-                        }
-                    } else {
-                        ascii = convert_ascii_vec(t, h->v.h_vector);
-                    }
-
-                    ascii2 = ascii;
-                    if (*ascii == '?') {
-                        wave_rgb_t cb;
-                        char *srch_for_color = strchr(ascii + 1, '?');
-                        if (srch_for_color) {
-                            *srch_for_color = 0;
-                            cb = XXX_get_gc_from_name(ascii + 1);
-                            if (cb.a != 0.0) {
-                                ascii2 = srch_for_color + 1;
-                                if (memcmp(&GLOBALS->rgb_gc.gc_back_wavewindow_c_1,
-                                           &GLOBALS->rgb_gc_white,
-                                           sizeof(wave_rgb_t))) {
-                                    if (!GLOBALS->black_and_white)
-                                        XXX_gdk_draw_rectangle(
-                                            GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                            cb,
-                                            TRUE,
-                                            _x0 + 1,
-                                            _y1 + 1,
-                                            width - 1,
-                                            (_y0 - 1) - (_y1 + 1) + 1);
-                                }
-                                GLOBALS->fill_in_smaller_rgb_areas_wavewindow_c_1 = 1;
-                            } else {
-                                *srch_for_color = '?'; /* replace name as color is a miss */
-                            }
-                        }
-                    }
-
-                    if ((_x1 >= GLOBALS->wavewidth) ||
-                        (font_engine_string_measure(GLOBALS->wavefont, ascii2) +
-                             GLOBALS->vector_padding <=
-                         width)) {
-                        XXX_font_engine_draw_string(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                                    GLOBALS->wavefont,
-                                                    &GLOBALS->rgb_gc.gc_value_wavewindow_c_1,
-                                                    _x0 + 2 + WAVE_CAIRO_050_OFFSET,
-                                                    ytext + WAVE_CAIRO_050_OFFSET,
-                                                    ascii2);
-                    } else {
-                        char *mod;
-
-                        mod = bsearch_trunc(ascii2, width - GLOBALS->vector_padding);
-                        if (mod) {
-                            *mod = '+';
-                            *(mod + 1) = 0;
-
-                            XXX_font_engine_draw_string(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                                        GLOBALS->wavefont,
-                                                        &GLOBALS->rgb_gc.gc_value_wavewindow_c_1,
-                                                        _x0 + 2 + WAVE_CAIRO_050_OFFSET,
-                                                        ytext + WAVE_CAIRO_050_OFFSET,
-                                                        ascii2);
-                        }
-                    }
-                } else if (GLOBALS->fill_in_smaller_rgb_areas_wavewindow_c_1) {
-                    /* char *ascii2; */ /* scan-build */
-
-                    if (h->flags & HIST_REAL) {
-                        if (!(h->flags & HIST_STRING)) {
-#ifdef WAVE_HAS_H_DOUBLE
-                            ascii = convert_ascii_real(t, &h->v.h_double);
-#else
-                            ascii = convert_ascii_real(t, (double *)h->v.h_vector);
-#endif
-                        } else {
-                            ascii = convert_ascii_string((char *)h->v.h_vector);
-                        }
-                    } else {
-                        ascii = convert_ascii_vec(t, h->v.h_vector);
-                    }
-
-                    /* ascii2 = ascii; */ /* scan-build */
-                    if (*ascii == '?') {
-                        wave_rgb_t cb;
-                        char *srch_for_color = strchr(ascii + 1, '?');
-                        if (srch_for_color) {
-                            *srch_for_color = 0;
-                            cb = XXX_get_gc_from_name(ascii + 1);
-                            if (cb.a != 0.0) {
-                                /* ascii2 =  srch_for_color + 1; */ /* scan-build */
-                                if (memcmp(&GLOBALS->rgb_gc.gc_back_wavewindow_c_1,
-                                           &GLOBALS->rgb_gc_white,
-                                           sizeof(wave_rgb_t))) {
-                                    if (!GLOBALS->black_and_white)
-                                        XXX_gdk_draw_rectangle(
-                                            GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                            cb,
-                                            TRUE,
-                                            _x0,
-                                            _y1 + 1,
-                                            width,
-                                            (_y0 - 1) - (_y1 + 1) + 1);
-                                }
-                            } else {
-                                *srch_for_color = '?'; /* replace name as color is a miss */
-                            }
-                        }
-                    }
-                }
-            }
-        } else {
-            newtime = (((gdouble)(_x1 + WAVE_OPT_SKIP)) * GLOBALS->nspx) +
-                      GLOBALS->tims.start /*+GLOBALS->shift_timebase*/; /* skip to next pixel */
-            h3 = bsearch_node(t->n.nd, newtime);
-            if (h3->time > h->time) {
-                h = h3;
-                lasttype = type;
-                continue;
-            }
-        }
-
-        if (ascii) {
-            free_2(ascii);
-            ascii = NULL;
-        }
-        h = h->next;
-        lasttype = type;
-    }
-
-    GLOBALS->color_active_in_filter = 0;
-
-    GLOBALS->tims.start += GLOBALS->shift_timebase;
-    GLOBALS->tims.end += GLOBALS->shift_timebase;
-}
-
-/********************************************************************************************************/
-
-static void draw_vptr_trace_analog(Trptr t, vptr v, int which, int num_extension)
-{
-    TimeType _x0, _x1, newtime;
-    int _y0, _y1, yu, liney, yt0, yt1;
-    TimeType tim, h2tim;
-    vptr h, h2, h3;
-    int endcnt = 0;
-    int type;
-    /* int lasttype=-1; */ /* scan-build */
-    wave_rgb_t c, ci;
-    wave_rgb_t cnan = GLOBALS->rgb_gc.gc_u_wavewindow_c_1;
-    wave_rgb_t cinf = GLOBALS->rgb_gc.gc_w_wavewindow_c_1;
-    wave_rgb_t cfixed;
-    double mynan = strtod("NaN", NULL);
-    double tmin = mynan, tmax = mynan, tv = 0.0, tv2;
-    gint rmargin;
-    int is_nan = 0, is_nan2 = 0, is_inf = 0, is_inf2 = 0;
-    int any_infs = 0, any_infp = 0, any_infm = 0;
-    int skipcnt = 0;
-
-    ci = GLOBALS->rgb_gc.gc_baseline_wavewindow_c_1;
-
-    h = v;
-    liney = ((which + 2 + num_extension) * GLOBALS->fontheight) - 2;
-    _y1 = ((which + 1) * GLOBALS->fontheight) + 2;
-    _y0 = liney - 2;
-    yu = (_y0 + _y1) / 2;
-
-    if (t->flags & TR_ANALOG_FULLSCALE) /* otherwise use dynamic */
-    {
-        if ((!t->minmax_valid) || (t->d_num_ext != num_extension)) {
-            h3 = t->n.vec->vectors[0];
-            for (;;) {
-                if (!h3)
-                    break;
-
-                if ((h3->time >= GLOBALS->tims.first) && (h3->time <= GLOBALS->tims.last)) {
-                    /* tv = mynan; */ /* scan-build */
-
-                    tv = convert_real(t, h3);
-                    if (!isnan(tv) && !isinf(tv)) {
-                        if (isnan(tmin) || tv < tmin)
-                            tmin = tv;
-                        if (isnan(tmax) || tv > tmax)
-                            tmax = tv;
-                    }
-                } else if (isinf(tv)) {
-                    any_infs = 1;
-
-                    if (tv > 0) {
-                        any_infp = 1;
-                    } else {
-                        any_infm = 1;
-                    }
-                }
-
-                h3 = h3->next;
-            }
-
-            if (isnan(tmin) || isnan(tmax))
-                tmin = tmax = 0;
-
-            if (any_infs) {
-                double tdelta = (tmax - tmin) * WAVE_INF_SCALING;
-
-                if (any_infp)
-                    tmax = tmax + tdelta;
-                if (any_infm)
-                    tmin = tmin - tdelta;
-            }
-
-            if ((tmax - tmin) < 1e-20) {
-                tmax = 1;
-                tmin -= 0.5 * (_y1 - _y0);
-            } else {
-                tmax = (_y1 - _y0) / (tmax - tmin);
-            }
-
-            t->minmax_valid = 1;
-            t->d_minval = tmin;
-            t->d_maxval = tmax;
-            t->d_num_ext = num_extension;
-        } else {
-            tmin = t->d_minval;
-            tmax = t->d_maxval;
-        }
-    } else {
-        h3 = h;
-        for (;;) {
-            if (!h3)
-                break;
-            tim = (h3->time);
-
-            if (tim > GLOBALS->tims.end) {
-                endcnt++;
-                if (endcnt == 2)
-                    break;
-            }
-            if (tim > GLOBALS->tims.last)
-                break;
-
-            _x0 = (tim - GLOBALS->tims.start) * GLOBALS->pxns;
-            if ((_x0 > GLOBALS->wavewidth) && (endcnt == 2)) {
-                break;
-            }
-
-            tv = convert_real(t, h3);
-            if (!isnan(tv) && !isinf(tv)) {
-                if (isnan(tmin) || tv < tmin)
-                    tmin = tv;
-                if (isnan(tmax) || tv > tmax)
-                    tmax = tv;
-            } else if (isinf(tv)) {
-                any_infs = 1;
-                if (tv > 0) {
-                    any_infp = 1;
-                } else {
-                    any_infm = 1;
-                }
-            }
-
-            h3 = h3->next;
-        }
-        if (isnan(tmin) || isnan(tmax))
-            tmin = tmax = 0;
-
-        if (any_infs) {
-            double tdelta = (tmax - tmin) * WAVE_INF_SCALING;
-
-            if (any_infp)
-                tmax = tmax + tdelta;
-            if (any_infm)
-                tmin = tmin - tdelta;
-        }
-
-        if ((tmax - tmin) < 1e-20) {
-            tmax = 1;
-            tmin -= 0.5 * (_y1 - _y0);
-        } else {
-            tmax = (_y1 - _y0) / (tmax - tmin);
-        }
-    }
-
-    if (GLOBALS->tims.last - GLOBALS->tims.start < GLOBALS->wavewidth) {
-        rmargin = (GLOBALS->tims.last - GLOBALS->tims.start) * GLOBALS->pxns;
-    } else {
-        rmargin = GLOBALS->wavewidth;
-    }
-
-    h3 = NULL;
-    for (;;) {
-        if (!h)
-            break;
-        tim = (h->time);
-        if ((tim > GLOBALS->tims.end) || (tim > GLOBALS->tims.last))
-            break;
-
-        _x0 = (tim - GLOBALS->tims.start) * GLOBALS->pxns;
-
-        /*
-        if(_x0<-1)
-                {
-                _x0=-1;
-                }
-                else
-        if(_x0>GLOBALS->wavewidth)
-                {
-                break;
-                }
-        */
-
-        h2 = h->next;
-        if (!h2)
-            break;
-        h2tim = tim = (h2->time);
-        if (tim > GLOBALS->tims.last)
-            tim = GLOBALS->tims.last;
-        /*	else if(tim>GLOBALS->tims.end+1) tim=GLOBALS->tims.end+1; */
-        _x1 = (tim - GLOBALS->tims.start) * GLOBALS->pxns;
-
-        /*
-        if(_x1<-1)
-                {
-                _x1=-1;
-                }
-                else
-        if(_x1>GLOBALS->wavewidth)
-                {
-                _x1=GLOBALS->wavewidth;
-                }
-        */
-        /* draw trans */
-        type = vtype2(t, h);
-        tv = convert_real(t, h);
-        tv2 = convert_real(t, h2);
-
-        if ((is_inf = isinf(tv))) {
-            if (tv < 0) {
-                yt0 = _y0;
-            } else {
-                yt0 = _y1;
-            }
-        } else if ((is_nan = isnan(tv))) {
-            yt0 = yu;
-        } else {
-            yt0 = _y0 + (tv - tmin) * tmax;
-        }
-
-        if ((is_inf2 = isinf(tv2))) {
-            if (tv2 < 0) {
-                yt1 = _y0;
-            } else {
-                yt1 = _y1;
-            }
-        } else if ((is_nan2 = isnan(tv2))) {
-            yt1 = yu;
-        } else {
-            yt1 = _y0 + (tv2 - tmin) * tmax;
-        }
-
-        if ((_x0 != _x1) ||
-            (skipcnt < GLOBALS->analog_redraw_skip_count)) /* lower number = better performance */
-        {
-            if (_x0 == _x1) {
-                skipcnt++;
-            } else {
-                skipcnt = 0;
-            }
-
-            if (type != AN_X) {
-                c = GLOBALS->rgb_gc.gc_vbox_wavewindow_c_1;
-            } else {
-                c = GLOBALS->rgb_gc.gc_x_wavewindow_c_1;
-            }
-
-            if (h->next) {
-                if (h->next->time > GLOBALS->max_time) {
-                    yt1 = yt0;
-                }
-            }
-
-            cfixed = is_inf ? cinf : c;
-            if ((is_nan2) && (h2tim > GLOBALS->max_time))
-                is_nan2 = 0;
-
-            /* clamp to top/bottom because of integer rounding errors */
-
-            if (yt0 < _y1)
-                yt0 = _y1;
-            else if (yt0 > _y0)
-                yt0 = _y0;
-
-            if (yt1 < _y1)
-                yt1 = _y1;
-            else if (yt1 > _y0)
-                yt1 = _y0;
-
-            /* clipping... */
-            {
-                int coords[4];
-                int rect[4];
-
-                if (_x0 < INT_MIN) {
-                    coords[0] = INT_MIN;
-                } else if (_x0 > INT_MAX) {
-                    coords[0] = INT_MAX;
-                } else {
-                    coords[0] = _x0;
-                }
-
-                if (_x1 < INT_MIN) {
-                    coords[2] = INT_MIN;
-                } else if (_x1 > INT_MAX) {
-                    coords[2] = INT_MAX;
-                } else {
-                    coords[2] = _x1;
-                }
-
-                coords[1] = yt0;
-                coords[3] = yt1;
-
-                rect[0] = -10;
-                rect[1] = _y1;
-                rect[2] = GLOBALS->wavewidth + 10;
-                rect[3] = _y0;
-
-                if ((t->flags & (TR_ANALOG_INTERPOLATED | TR_ANALOG_STEP)) != TR_ANALOG_STEP) {
-                    wave_lineclip(coords, rect);
-                } else {
-                    if (coords[0] < rect[0])
-                        coords[0] = rect[0];
-                    if (coords[2] < rect[0])
-                        coords[2] = rect[0];
-
-                    if (coords[0] > rect[2])
-                        coords[0] = rect[2];
-                    if (coords[2] > rect[2])
-                        coords[2] = rect[2];
-
-                    if (coords[1] < rect[1])
-                        coords[1] = rect[1];
-                    if (coords[3] < rect[1])
-                        coords[3] = rect[1];
-
-                    if (coords[1] > rect[3])
-                        coords[1] = rect[3];
-                    if (coords[3] > rect[3])
-                        coords[3] = rect[3];
-                }
-
-                _x0 = coords[0];
-                yt0 = coords[1];
-                _x1 = coords[2];
-                yt1 = coords[3];
-            }
-            /* ...clipping */
-
-            if (is_nan || is_nan2) {
-                if (is_nan) {
-                    XXX_gdk_draw_rectangle(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                           cnan,
-                                           TRUE,
-                                           _x0,
-                                           _y1,
-                                           _x1 - _x0,
-                                           _y0 - _y1);
-
-                    if ((t->flags & (TR_ANALOG_INTERPOLATED | TR_ANALOG_STEP)) ==
-                        (TR_ANALOG_INTERPOLATED | TR_ANALOG_STEP)) {
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          ci,
-                                          _x1 - 1,
-                                          yt1,
-                                          _x1 + 1,
-                                          yt1);
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          ci,
-                                          _x1,
-                                          yt1 - 1,
-                                          _x1,
-                                          yt1 + 1);
-
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          ci,
-                                          _x0 - 1,
-                                          _y0,
-                                          _x0 + 1,
-                                          _y0);
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          ci,
-                                          _x0,
-                                          _y0 - 1,
-                                          _x0,
-                                          _y0 + 1);
-
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          ci,
-                                          _x0 - 1,
-                                          _y1,
-                                          _x0 + 1,
-                                          _y1);
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          ci,
-                                          _x0,
-                                          _y1 - 1,
-                                          _x0,
-                                          _y1 + 1);
-                    }
-                }
-                if (is_nan2) {
-                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                      cfixed,
-                                      _x0,
-                                      yt0,
-                                      _x1,
-                                      yt0);
-
-                    if ((t->flags & (TR_ANALOG_INTERPOLATED | TR_ANALOG_STEP)) ==
-                        (TR_ANALOG_INTERPOLATED | TR_ANALOG_STEP)) {
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          cnan,
-                                          _x1,
-                                          yt1,
-                                          _x1,
-                                          yt0);
-
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          ci,
-                                          _x1 - 1,
-                                          _y0,
-                                          _x1 + 1,
-                                          _y0);
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          ci,
-                                          _x1,
-                                          _y0 - 1,
-                                          _x1,
-                                          _y0 + 1);
-
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          ci,
-                                          _x1 - 1,
-                                          _y1,
-                                          _x1 + 1,
-                                          _y1);
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          ci,
-                                          _x1,
-                                          _y1 - 1,
-                                          _x1,
-                                          _y1 + 1);
-                    }
-                }
-            } else if ((t->flags & TR_ANALOG_INTERPOLATED) && !is_inf && !is_inf2) {
-                if (t->flags & TR_ANALOG_STEP) {
-                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                      ci,
-                                      _x0 - 1,
-                                      yt0,
-                                      _x0 + 1,
-                                      yt0);
-                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                      ci,
-                                      _x0,
-                                      yt0 - 1,
-                                      _x0,
-                                      yt0 + 1);
-                }
-
-                if (rmargin != GLOBALS->wavewidth) /* the window is clipped in postscript */
-                {
-                    if ((yt0 == yt1) && ((_x0 > _x1) || (_x0 < 0))) {
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          cfixed,
-                                          0,
-                                          yt0,
-                                          _x1,
-                                          yt1);
-                    } else {
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          cfixed,
-                                          _x0,
-                                          yt0,
-                                          _x1,
-                                          yt1);
-                    }
-                } else {
-                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                      cfixed,
-                                      _x0,
-                                      yt0,
-                                      _x1,
-                                      yt1);
-                }
-            } else
-            /* if(t->flags & TR_ANALOG_STEP) */
-            {
-                XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                  cfixed,
-                                  _x0,
-                                  yt0,
-                                  _x1,
-                                  yt0);
-
-                if (is_inf2)
-                    cfixed = cinf;
-                XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                  cfixed,
-                                  _x1,
-                                  yt0,
-                                  _x1,
-                                  yt1);
-
-                if ((t->flags & (TR_ANALOG_INTERPOLATED | TR_ANALOG_STEP)) ==
-                    (TR_ANALOG_INTERPOLATED | TR_ANALOG_STEP)) {
-                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                      ci,
-                                      _x0 - 1,
-                                      yt0,
-                                      _x0 + 1,
-                                      yt0);
-                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                      ci,
-                                      _x0,
-                                      yt0 - 1,
-                                      _x0,
-                                      yt0 + 1);
-                }
-            }
-        } else {
-            newtime = (((gdouble)(_x1 + WAVE_OPT_SKIP)) * GLOBALS->nspx) +
-                      GLOBALS->tims.start /*+GLOBALS->shift_timebase*/; /* skip to next pixel */
-            h3 = bsearch_vector(t->n.vec, newtime);
-            if (h3->time > h->time) {
-                h = h3;
-                /* lasttype=type; */
-                continue;
-            }
-        }
-
-        h = h->next;
-        /* lasttype=type; */
-    }
-
-    GLOBALS->tims.start += GLOBALS->shift_timebase;
-    GLOBALS->tims.end += GLOBALS->shift_timebase;
-}
-
-/*
- * draw vector traces
- */
-static void draw_vptr_trace(Trptr t, vptr v, int which)
-{
-    TimeType _x0, _x1, newtime, width;
-    int _y0, _y1, yu, liney, ytext;
-    TimeType tim /* , h2tim */; /* scan-build */
-    vptr h, h2, h3;
-    char *ascii = NULL;
-    int type;
-    int lasttype = -1;
-    wave_rgb_t c;
-
-    GLOBALS->tims.start -= GLOBALS->shift_timebase;
-    GLOBALS->tims.end -= GLOBALS->shift_timebase;
-
-    liney = ((which + 2) * GLOBALS->fontheight) - 2;
-    _y1 = ((which + 1) * GLOBALS->fontheight) + 2;
-    _y0 = liney - 2;
-    yu = (_y0 + _y1) / 2;
-    ytext = yu - (GLOBALS->wavefont->ascent / 2) + GLOBALS->wavefont->ascent;
-
-    if ((GLOBALS->highlight_wavewindow) && (t) && (t->flags & TR_HIGHLIGHT) &&
-        (!GLOBALS->black_and_white)) {
-        Trptr tn = GiveNextTrace(t);
-        if ((t->flags & TR_ANALOGMASK) && (tn) && (tn->flags & TR_ANALOG_BLANK_STRETCH)) {
-            XXX_gdk_draw_rectangle(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                   GLOBALS->rgb_gc.gc_grid_wavewindow_c_1,
-                                   TRUE,
-                                   0,
-                                   liney - GLOBALS->fontheight,
-                                   GLOBALS->wavewidth,
-                                   GLOBALS->fontheight);
-        } else {
-            XXX_gdk_draw_rectangle(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                   GLOBALS->rgb_gc.gc_grid_wavewindow_c_1,
-                                   TRUE,
-                                   0,
-                                   liney - GLOBALS->fontheight,
-                                   GLOBALS->wavewidth,
-                                   GLOBALS->fontheight);
-        }
-    } else if ((GLOBALS->display_grid) && (GLOBALS->enable_horiz_grid)) {
-        Trptr tn = GiveNextTrace(t);
-        if ((t->flags & TR_ANALOGMASK) && (tn) && (tn->flags & TR_ANALOG_BLANK_STRETCH)) {
-        } else {
-            XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                              GLOBALS->rgb_gc.gc_grid_wavewindow_c_1,
-                              (GLOBALS->tims.start < GLOBALS->tims.first)
-                                  ? (GLOBALS->tims.first - GLOBALS->tims.start) * GLOBALS->pxns
-                                  : 0,
-                              liney,
-                              (GLOBALS->tims.last <= GLOBALS->tims.end)
-                                  ? (GLOBALS->tims.last - GLOBALS->tims.start) * GLOBALS->pxns
-                                  : GLOBALS->wavewidth - 1,
-                              liney);
-        }
-    }
-
-    h = v;
-    /* obsolete:
-    if(t->flags & TR_TTRANSLATED)
-            {
-            traverse_vector_nodes(t);
-            h=v=bsearch_vector(t->n.vec, GLOBALS->tims.start);
-            }
-            else
-            {
-            h=v;
-            }
-    */
-
-    if (t->flags & TR_ANALOGMASK) {
-        Trptr te = GiveNextTrace(t);
-        int ext = 0;
-
-        while (te) {
-            if (te->flags & TR_ANALOG_BLANK_STRETCH) {
-                ext++;
-                te = GiveNextTrace(te);
-            } else {
-                break;
-            }
-        }
-
-        if ((ext) && (GLOBALS->highlight_wavewindow) && (t) && (t->flags & TR_HIGHLIGHT) &&
-            (!GLOBALS->black_and_white)) {
-            XXX_gdk_draw_rectangle(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                   GLOBALS->rgb_gc.gc_grid_wavewindow_c_1,
-                                   TRUE,
-                                   0,
-                                   liney,
-                                   GLOBALS->wavewidth,
-                                   GLOBALS->fontheight * ext);
-        }
-
-        draw_vptr_trace_analog(t, v, which, ext);
-
-        GLOBALS->tims.start += GLOBALS->shift_timebase;
-        GLOBALS->tims.end += GLOBALS->shift_timebase;
-        return;
-    }
-
-    GLOBALS->color_active_in_filter = 1;
-
-    for (;;) {
-        if (!h)
-            break;
-        tim = (h->time);
-        if ((tim > GLOBALS->tims.end) || (tim > GLOBALS->tims.last))
-            break;
-
-        _x0 = (tim - GLOBALS->tims.start) * GLOBALS->pxns;
-        if (_x0 < -1) {
-            _x0 = -1;
-        } else if (_x0 > GLOBALS->wavewidth) {
-            break;
-        }
-
-        h2 = h->next;
-        if (!h2)
-            break;
-        /* h2tim= */ tim = (h2->time); /* scan-build */
-        if (tim > GLOBALS->tims.last)
-            tim = GLOBALS->tims.last;
-        else if (tim > GLOBALS->tims.end + 1)
-            tim = GLOBALS->tims.end + 1;
-        _x1 = (tim - GLOBALS->tims.start) * GLOBALS->pxns;
-        if (_x1 < -1) {
-            _x1 = -1;
-        } else if (_x1 > GLOBALS->wavewidth) {
-            _x1 = GLOBALS->wavewidth;
-        }
-
-        /* draw trans */
-        type = vtype2(t, h);
-
-        if (_x0 > -1) {
-            wave_rgb_t gltype, gtype;
-
-            switch (lasttype) {
-                case AN_X:
-                    gltype = GLOBALS->rgb_gc.gc_x_wavewindow_c_1;
-                    break;
-                case AN_U:
-                    gltype = GLOBALS->rgb_gc.gc_u_wavewindow_c_1;
-                    break;
-                default:
-                    gltype = GLOBALS->rgb_gc.gc_vtrans_wavewindow_c_1;
-                    break;
-            }
-            switch (type) {
-                case AN_X:
-                    gtype = GLOBALS->rgb_gc.gc_x_wavewindow_c_1;
-                    break;
-                case AN_U:
-                    gtype = GLOBALS->rgb_gc.gc_u_wavewindow_c_1;
-                    break;
-                default:
-                    gtype = GLOBALS->rgb_gc.gc_vtrans_wavewindow_c_1;
-                    break;
-            }
-
-            if (GLOBALS->use_roundcaps) {
-                if (type == AN_Z) {
-                    if (lasttype != -1) {
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          gltype,
-                                          _x0 - 1,
-                                          _y0,
-                                          _x0,
-                                          yu);
-                        if (lasttype != AN_0)
-                            XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                              gltype,
-                                              _x0,
-                                              yu,
-                                              _x0 - 1,
-                                              _y1);
-                    }
-                } else if (lasttype == AN_Z) {
-                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                      gtype,
-                                      _x0 + 1,
-                                      _y0,
-                                      _x0,
-                                      yu);
-                    if (type != AN_0)
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          gtype,
-                                          _x0,
-                                          yu,
-                                          _x0 + 1,
-                                          _y1);
-                } else {
-                    if (lasttype != type) {
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          gltype,
-                                          _x0 - 1,
-                                          _y0,
-                                          _x0,
-                                          yu);
-                        if (lasttype != AN_0)
-                            XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                              gltype,
-                                              _x0,
-                                              yu,
-                                              _x0 - 1,
-                                              _y1);
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          gtype,
-                                          _x0 + 1,
-                                          _y0,
-                                          _x0,
-                                          yu);
-                        if (type != AN_0)
-                            XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                              gtype,
-                                              _x0,
-                                              yu,
-                                              _x0 + 1,
-                                              _y1);
-                    } else {
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          gtype,
-                                          _x0 - 2,
-                                          _y0,
-                                          _x0 + 2,
-                                          _y1);
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          gtype,
-                                          _x0 + 2,
-                                          _y0,
-                                          _x0 - 2,
-                                          _y1);
-                    }
-                }
-            } else {
-                XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1, gtype, _x0, _y0, _x0, _y1);
-            }
-        }
-
-        if (_x0 != _x1) {
-            if (type == AN_Z) {
-                if (GLOBALS->use_roundcaps) {
-                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                      GLOBALS->rgb_gc.gc_mid_wavewindow_c_1,
-                                      _x0 + 1,
-                                      yu,
-                                      _x1 - 1,
-                                      yu);
-                } else {
-                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                      GLOBALS->rgb_gc.gc_mid_wavewindow_c_1,
-                                      _x0,
-                                      yu,
-                                      _x1,
-                                      yu);
-                }
-            } else {
-                if ((type != AN_X) && (type != AN_U)) {
-                    c = GLOBALS->rgb_gc.gc_vbox_wavewindow_c_1;
-                } else {
-                    c = GLOBALS->rgb_gc.gc_x_wavewindow_c_1;
-                }
-
-                if (GLOBALS->use_roundcaps) {
-                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                      c,
-                                      _x0 + 2,
-                                      _y0,
-                                      _x1 - 2,
-                                      _y0);
-                    if (type != AN_0)
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          c,
-                                          _x0 + 2,
-                                          _y1,
-                                          _x1 - 2,
-                                          _y1);
-                    if (type == AN_1)
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          c,
-                                          _x0 + 2,
-                                          _y1 + 1,
-                                          _x1 - 2,
-                                          _y1 + 1);
-                } else {
-                    XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1, c, _x0, _y0, _x1, _y0);
-                    if (type != AN_0)
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          c,
-                                          _x0,
-                                          _y1,
-                                          _x1,
-                                          _y1);
-                    if (type == AN_1)
-                        XXX_gdk_draw_line(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                          c,
-                                          _x0,
-                                          _y1 + 1,
-                                          _x1,
-                                          _y1 + 1);
-                }
-
-                if (_x0 < 0)
-                    _x0 = 0; /* fixup left margin */
-
-                if ((width = _x1 - _x0) > GLOBALS->vector_padding) {
-                    char *ascii2;
-
-                    ascii = convert_ascii(t, h);
-
-                    ascii2 = ascii;
-                    if (*ascii == '?') {
-                        wave_rgb_t cb;
-                        char *srch_for_color = strchr(ascii + 1, '?');
-                        if (srch_for_color) {
-                            *srch_for_color = 0;
-                            cb = XXX_get_gc_from_name(ascii + 1);
-                            if (cb.a != 0.0) {
-                                ascii2 = srch_for_color + 1;
-                                if (!GLOBALS->black_and_white)
-                                    XXX_gdk_draw_rectangle(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                                           cb,
-                                                           TRUE,
-                                                           _x0 + 1,
-                                                           _y1 + 1,
-                                                           width - 1,
-                                                           (_y0 - 1) - (_y1 + 1) + 1);
-                                GLOBALS->fill_in_smaller_rgb_areas_wavewindow_c_1 = 1;
-                            } else {
-                                *srch_for_color = '?'; /* replace name as color is a miss */
-                            }
-                        }
-                    }
-
-                    if ((_x1 >= GLOBALS->wavewidth) ||
-                        (font_engine_string_measure(GLOBALS->wavefont, ascii2) +
-                             GLOBALS->vector_padding <=
-                         width)) {
-                        XXX_font_engine_draw_string(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                                    GLOBALS->wavefont,
-                                                    &GLOBALS->rgb_gc.gc_value_wavewindow_c_1,
-                                                    _x0 + 2,
-                                                    ytext,
-                                                    ascii2);
-                    } else {
-                        char *mod;
-
-                        mod = bsearch_trunc(ascii2, width - GLOBALS->vector_padding);
-                        if (mod) {
-                            *mod = '+';
-                            *(mod + 1) = 0;
-
-                            XXX_font_engine_draw_string(GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                                        GLOBALS->wavefont,
-                                                        &GLOBALS->rgb_gc.gc_value_wavewindow_c_1,
-                                                        _x0 + 2,
-                                                        ytext,
-                                                        ascii2);
-                        }
-                    }
-
-                } else if (GLOBALS->fill_in_smaller_rgb_areas_wavewindow_c_1) {
-                    /* char *ascii2; */ /* scan-build */
-
-                    ascii = convert_ascii(t, h);
-
-                    /* ascii2 = ascii; */ /* scan-build */
-                    if (*ascii == '?') {
-                        wave_rgb_t cb;
-                        char *srch_for_color = strchr(ascii + 1, '?');
-                        if (srch_for_color) {
-                            *srch_for_color = 0;
-                            cb = XXX_get_gc_from_name(ascii + 1);
-                            if (cb.a != 0.0) {
-                                /* ascii2 =  srch_for_color + 1; */
-                                if (memcmp(&GLOBALS->rgb_gc.gc_back_wavewindow_c_1,
-                                           &GLOBALS->rgb_gc_white,
-                                           sizeof(wave_rgb_t))) {
-                                    if (!GLOBALS->black_and_white)
-                                        XXX_gdk_draw_rectangle(
-                                            GLOBALS->cr_wavepixmap_wavewindow_c_1,
-                                            cb,
-                                            TRUE,
-                                            _x0,
-                                            _y1 + 1,
-                                            width,
-                                            (_y0 - 1) - (_y1 + 1) + 1);
-                                }
-                            } else {
-                                *srch_for_color = '?'; /* replace name as color is a miss */
-                            }
-                        }
-                    }
-                }
-            }
-        } else {
-            newtime = (((gdouble)(_x1 + WAVE_OPT_SKIP)) * GLOBALS->nspx) +
-                      GLOBALS->tims.start /*+GLOBALS->shift_timebase*/; /* skip to next pixel */
-            h3 = bsearch_vector(t->n.vec, newtime);
-            if (h3->time > h->time) {
-                h = h3;
-                lasttype = type;
-                continue;
-            }
-        }
-
-        if (ascii) {
-            free_2(ascii);
-            ascii = NULL;
-        }
-        lasttype = type;
-        h = h->next;
-    }
-
-    GLOBALS->color_active_in_filter = 0;
-
-    GLOBALS->tims.start += GLOBALS->shift_timebase;
-    GLOBALS->tims.end += GLOBALS->shift_timebase;
 }
